### 2022-06-25 #4500 @Kevin-Mattheus-Moerman
@derNarr can you check those potentially missing DOIs? :point_up: You can call `@editorialbot generate pdf` to update the paper, and call `@editorialbot check references` to check those DOIs again. Thanks
