### 2022-06-25 #4500 @derNarr
I will check the DOIs on Monday. Thank you for finding and searching for reviewers!

Am 25. Juni 2022 17:23:56 MESZ schrieb Kevin Mattheus Moerman ***@***.***>:
***@***.*** can you check those potentially missing DOIs? :point_up: You can call ***@***.*** generate pdf` to update the paper, and call ***@***.*** check references` to check those DOIs again. Thanks
>
>--
>Reply to this email directly or view it on GitHub:
>https://github.com/openjournals/joss-reviews/issues/4500#issuecomment-1166309658
>You are receiving this because you were mentioned.
>
>Message ID: ***@***.***>
