### 2022-06-26 #4500 @jinhangjiang
Thank you for reaching out! I’m happy to review if there is some time left. Please let me know when the ddl is so that I can let you know my final decision.

Sent from my iPhone

> On Jun 25, 2022, at 12:14, Øystein Sørensen ***@***.***> wrote:
>
> ﻿
> 👋 @frankier @jinhangjiang @VenkteshV would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
