### 2022-06-26 #4500 @jinhangjiang
﻿Yeah. That definitely works for me. Please add me. I will follow up soon. Thank you again for reaching out to me.

Sent from my iPhone

> On Jun 26, 2022, at 02:39, Øystein Sørensen ***@***.***> wrote:
> ﻿
> @jinhangjiang thanks, there are no hard deadlines here, but if you accept, we would appreciate if you could start the review within three weeks. Is that ok for you?
>
