### 2022-06-27 #4515 @editorialbot
[REVIEW]: pyndl: Naïve Discriminative Learning in Python
**Submitting author:** @derNarr (<a href="http://orcid.org/0000-0002-1178-7932">Konstantin Sering</a>)
**Repository:** https://github.com/quantling/pyndl
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.1
