+++ paper/paper.bib
@@ -9,7 +9,8 @@
   title={Adaptive switching circuits},
   author={Widrow, Bernard and Hoff, Marcian E},
   year={1960},
-  institution={Stanford Univ Ca Stanford Electronics Labs}
+  institution={Stanford Univ Ca Stanford Electronics Labs},
+  doi = {10.21236/ad0241531}
 }
 
 @article{Danks_2003,
@@ -163,14 +164,15 @@
 }
 
 @article{Rescorla1988PavlovianCI,
-  title={Pavlovian conditioning. It's not what you think it is.},
+  title={Pavlovian conditioning: It's not what you think it is.},
   author={Robert A Rescorla},
   journal={The American psychologist},
   year={1988},
-  volume={43 3},
-  pages={
-          151-60
-        }
+  volume={43}
+  number={3},
+  pages={151--60},
+  publisher={American Psychological Association},
+  doi={10.1037/0003-066x.43.3.151}
 }
 
 @article{Tomaschek:Duran:2019,
