### 2022-06-27 ebfc2aa
fixes missing comma in bib file
