+++ paper/paper.bib
@@ -168,7 +168,7 @@
   author={Robert A Rescorla},
   journal={The American psychologist},
   year={1988},
-  volume={43}
+  volume={43},
   number={3},
   pages={151--60},
   publisher={American Psychological Association},
