### 2022-06-27 #4515 @derNarr
Added missing DOIs and checked that they lead to the right reference.
