### 2022-07-05 quantling/pyndl#211 @frankier
Installation instructions do not specify Python headers are needed
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4515).

The installation instructions do not specify that a successful installation requires the Python C headers (these are included in  python-dev in Debian for example).
