### 2022-07-05 quantling/pyndl#212 @frankier
Possibility of adding a wheel to PyPI
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4515).

If you add wheels to PyPI, then the current installation instructions will work as long as the user is using an architecture or a Python version which a wheel has been built for. In case they are not, pip will fall back to the source distribution, so it may still be worth adding that the Python C headers may have to be installed in this case.

Here is the documentation for building a wheel: https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#wheels
