### 2022-07-05 quantling/pyndl#213 @frankier
Non-file-based (in-memory/streaming) usage?
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4515). However it goes a bit beyond the remit, so feel free to treat this issue especially as "take it or leave it".

The high level/documented API of this module has an interface with a fixed pipeline where most of the "payload data" is deserialised/serialised to files. This interface may be convenient for some workflows, but other users may have their own workflows or may already have their data in a particular format, and might want to put it into the correct format "in memory" either as a single chunk or as multiple chunks using a generator.

While it is the case that ndl can already take a generator rather than an events files, it simply serialises the generator and reloads it. It would be nice to restructure things so there are two strata:

 * File based api, which is implemented in terms of
 * Generator/in-memory API
