### 2022-07-06 quantling/pyndl#212 @derNarr
Thank you for this suggestion.

At the moment building a `wheel` or even an `sdist` with the `build` package, does not work with the following commands:

```
python3 -m build --wheel
python3 -m build --sdist
```

It throughs the following error:

```
(...)
  File "setup.py", line 12, in finalize_options
    __builtins__.__NUMPY_SETUP__ = False
AttributeError: 'dict' object has no attribute '__NUMPY_SETUP__'
```

which suggests that either setup.py is incomplete or some dependencies are not installed correctly with `build`.

Our (old) way of building the package still works, but the wheel, I can build that way has the wrong naming scheme, and I don't want to "fix" it by hand.

```
python setup.py bdist_wheel
python setup.py sdist
```

It might still be good to investigate this further and fix/change things so that we can build `pyndl` with `build`.
