### 2022-07-06 quantling/pyndl#211 @derNarr
I always felt they come with a proper `cython` installation but it does not hurt to add them into our documentation as well. I will think about where to place this information best. If you have any suggestion, where you would expect this, let me know as well. Thanks!

Here? https://pyndl.readthedocs.io/en/latest/installation.html
