### 2022-07-06 quantling/pyndl#212 @frankier
I can recommend that you use an alternative approach to the current numpy bootstrap to avoid this error. Please see this answer: https://stackoverflow.com/a/68038058

To elaborate a bit more, the currently advocated approach is to move from having things in `setup.py` to configuration files, namely `pyproject.toml` and `setup.cfg`. The advantage of the configuration file approach is no package code is executed at installation time, which leads to various benefits like faster dependency resolution and reduces the possibility of bad interactions with new Python package management software.
