### 2022-07-06 quantling/pyndl#211 @frankier
> I always felt they come with a proper cython installation

I don't think so, but I'm not 100% on that. Probably if you install Cython through the OS package manager it will pull in the headers, but I don't think Python headers have ever been packaged through PyPI (maybe they should be -- but it would also probably involve somehow hooking into the build process so the path would be passed through).

I did get an actual error about missing header files during installation since I happened to not have them installed for the version of Python I tested with. The normal instructions for Debian -like distributions is `build-essential python3-dev`. There may be some instructions somewhere you can link to or reuse e.g. https://py-generic-project.readthedocs.io/en/latest/installing.html -- although that guide seems to miss out that you (?maybe?) need to install Visual C++ on Windows. For most Python programmers this is old hat, but the audience for this package could include corpus linguists who may be Python beginners. Also it's more common for people to not have them installed now there are so many wheels available.

> Here? https://pyndl.readthedocs.io/en/latest/installation.html

Yes you can put the full instructions there and then note in the "Installation" section of the README something like that it is "assuming you have already installed the requirement to compile Python extensions" and direct to the full instructions. If you upload wheels you could omit this in the README since usually people will be able to install the wheel (but there are various reasons they may have to fall back to a source installation).
