### 2022-07-18 #4515 @jinhangjiang
Hi @osorensen, I have finished the first round of reviews. It looks like a mature, solid, and domain-dependent python package to me. The authors are able to provide sufficient documents to help users to understand what it can achieve. The example codes are reproducible from my end.

I only have two questions for the authors (@derNarr):

- First, based on what I read, the API only takes the data in a specific format. And it looks hard to scale up. Is it a normal convention in this field? Since one of the claimed contributions is to help study large corpora, it would be nice to explain a bit how well it can work with large text data.

- Second, since it can work with non-English text, it would be nice to give some showcases. Also, I was not able to locate a list of supported languages.

Please let me know what the next step is. Thanks!
