### 2022-07-18 quantling/pyndl#214 @derNarr
Add example with non-ascii characters to demonstracte Unicode support
As mentioned by @jinhangjiang in the JOSS review, it would be nice to add a short example with non-ascii characters to demonstrate the use of Unicode (utf-8; utf-16).
