### 2022-07-18 #4515 @derNarr
>     * First, based on what I read, the API only takes the data in a specific format. And it looks hard to scale up. Is it a normal convention in this field? Since one of the claimed contributions is to help study large corpora, it would be nice to explain a bit how well it can work with large text data.

The ``ndl.ndl`` learner expects the data to be in a tab-separated format where in the first column cues (inputs) are defined, where the only special delimiting character is the underscore ``_``. In the second column the outcomes (outputs) are defined in the same way. This tab-separated file can / should be gunzipped (deflate compressed). As you can continue the learning, learning events can be split up to multiple event files.

The advantage and the origin of this event file format is that it is extremely transparent to the researcher, easy to inspect and that the time course of the learning events are defined by going from top to bottom in the file. As the research in our group was always concerned with the ordering of learning events and found some evidence that a naturalistic learning order is superior to random learning in modeling human reaction times. Naturalistic learning events usually have the same cue (or outcome) reappear in one phase of the learning a lot, whereas it might not be present at all present in the rest of the learning events. An example would be the cue/outcome "pyndl", which nearly only appears in this discussion and in research that uses this package, but you will not be exposed tot the word "pyndl" randomly on the Internet.

To be honest, the data format was in use before I joined the group and wrote pyndl to compute the weight matrices from this data format. But I think, it is extremely easy, allows for transparent use, should be accessible and readable in 20 to 50 years from now on, is around for at least 10 years. Therefore, it has many good features and is only slightly bloated. With ``pyndl.corpus`` there exists some code to create event files from different sources like OpenSubtiltes files or simple text files.

The data format runs with 100,000,000 (100 million) learning events, and I don't see a reason, why it not should scale up further. Compared to image, video and audio, text is small. @jinhangjiang If you still have the concern, can you specify why you believe it would not scale up? What data do you have in mind with which sizes?

>     * Second, since it can work with non-English text, it would be nice to give some showcases. Also, I was not able to locate a list of supported languages.

Everything that can be encoded into letters with Unicode code points (utf-8, utf-16, ...) can be computed with ``pyndl``. You are right that we should add a small example that uses Chinese or Greek or Taiwanese symbols. I created an issue for that in pyndl https://github.com/quantling/pyndl/issues/214

We stress this point so much in the paper, as with the older implementations in the R programming language we regularly ran into problems with the Unicode processing, which was done in with ``libicu``, and therefore it was a big deal for us to be able to run the learning on all the languages we wanted. The only thing we still needed to do is convert texts to utf-8 or read them into python 3.

@jinhangjiang thank you for your questions and your review! Did I answered you questions? Do you suggest that we should change something or have other ideas of improvements? Let me know as well, if I should elaborate on one end or another.
