### 2022-07-18 #4515 @jinhangjiang
@derNarr Thank you for answering my questions throughout. I really appreciate the details you shared. When I said I thought it was hard to scale up, I was thinking about how to prepare the inputs in something like pandas/spark dataframe as I am more familiar with those frameworks. What you have explained makes sense to me. As you have experimented with 100 million learning events, I would suggest sharing some examples of estimated execution time with corresponding data volume. If I were about to use the package in my research, I would be interested in knowing this information in advance.

Looking forward to the small examples in different languages such as Chinese. It would be very interesting to see how it works with characters instead of letters.

Thank you for all your work and clarifications. Best.
