### 2022-08-01 quantling/pyndl#215 @frankier
Public benchmarks
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4515). I am following up on a point made by @jinhangjiang in their review.

It would be good to include benchmarks somewhere where people evaluating this library can easily see them.

In its simplest form this could be a simple run time on some moderately sized corpus and would include a rough specification of the benchmarking machine. Better would be to include a direct comparison with one or more of the comparable R libraries. At the top end you could a plot with many runs with increasing corpus size across different libraries.

Ideally the scripts to run all benchmarks would be included in the repository so others can rerun/verify/add more comparisons.

Since speed is mentioned in the JOSS article, it would be good if something in this direction could also be included there which would help to make the publication more precise.
