### 2022-08-01 quantling/pyndl#211 @frankier
This appears to have come up before https://github.com/quantling/pyndl/issues/187 (I did not find this issue when searching previously since it's a bit vague)
