### 2022-08-01 quantling/pyndl#213 @frankier
The normal way of handling a generic resettable stream like this in Python would be to provide multiple implementations which follow some kind of reader interface. This could be more functional style as a function which returns an iterable which is recalled every time we need to reset it, or it could be a implemented using a class (similar to file-like objects) that has `reset()` and `next()` methods. I think in terms of working with multiprocessing, the latter is likely to be a smoother experience. An in-memory interface and pyndl event format interface could be provided, and then the documentation could give details of how users can make their own reader if they want (noting that it will be used in a multiprocessing context).

This issue may be related: https://github.com/quantling/pyndl/issues/130
