### 2022-08-01 quantling/pyndl#217 @frankier
More conventional documentation TOC
I recommend reorganizing the documentation table of contents to match commonly encountered headings

"Naive Discriminative Learning" => "Background" (other background sections could go here too if need be)
"More Examples" => "Usage examples"
"Package overview" => "API documentation"
"Miscellaneous" => Something less vague, at least "Tips & Tricks", but ideally things in here would be properly categorized elsewhere (e.g. have a "Language interoperation" section)
"About" => Something more specific, perhaps "Credits"?
