### 2022-08-01 quantling/pyndl#219 @frankier
Quickstart section could more clearly outline purpose of examples and motivation behind different data preparation procedures
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4515).

The quickstart section could more clearly outline the purpose of the examples and the motivation behind the different data preparation procedures.

At a high level, I see there being 4 layers/strata in the project from the lowest:

 1. The NDL rule as independent of the file format used for the events and the weights matrix (does not seem to be directly exposed/documented -- see the comments of @jinhangjiang and https://github.com/quantling/pyndl/issues/213
    * Exposing/documenting this would be a "nice to have" and could create appeal to those outside the cognitive linguistic audience
 2. The NDL rule as implemented in terms of the library's custom file formats
    * This is introduced in the documentation through a word form-semantics association (meaning recognition) example
 3. Event file generation routines which can convert "END OF DOCUMENT" delimited corpora files into
    * This is introduced in the documentation between a bigram-word form association (form recognition) example
 4. Convenience functions to convert from other corpora formats (currently only OpenSubtitles) to stratum 3

In the documentation too, the strata could be sign-posted earlier, so as to prevent confusion. Stratum 2 and 3 are rather easy to get confused between at the moment. https://pyndl.readthedocs.io/en/latest/quickstart.html#correct-data-format deals with both. Here, the domain-purpose of the example datasets should be explained at the beginning of each example. Additionally, Strata 2 and 3 should probably be introduced briefly at the beginning of the whole before going into the examples.

"Correct Data Format" => "Data Preparation"
