### 2022-08-01 quantling/pyndl#213 @frankier
At a high level, I see there being 4 layers/strata in the project from the lowest:

 1. The NDL rule as independent of the file format used for the events and the weights matrix (does not seem to be directly exposed/documented -- see the comments of @jinhangjiang and https://github.com/quantling/pyndl/issues/213
   * Exposing/documenting this would be a "nice to have" and could create appeal to those outside the cognitive linguistic audience
 2. The NDL rule as implemented in terms of the library's custom file formats
   * This is introduced in the documentation through a word form-semantics association (meaning recognition) example
 3. Event file generation routines which can convert "END OF DOCUMENT" delimited corpora files into
   * This is introduced in the documentation between a bigram-word form association (form recognition) example
 4. Convenience functions to convert from other corpora formats (currently only OpenSubtitles) to stratum 3

Exposing and documenting Stratum 1 may be a good longer term goal.
