### 2022-08-01 quantling/pyndl#220 @frankier
create_event_file does not deal with accented characters(?)
### What you were trying to do

Run the following code:

```
from pyndl import preprocess

CORPUS = """
mitä
mikä
minä
miksi
""".strip()

with open("/tmp/corpus.txt", "w") as f:
    f.write(CORPUS)

if os.path.exists("/tmp/event.tab.gz"):
    os.remove("/tmp/event.tab.gz")

preprocess.create_event_file(
    corpus_file="/tmp/corpus.txt",
    event_file="/tmp/event.tab.gz",
    context_structure='document',
    event_structure='consecutive_words',
    event_options=(1, ),
    cue_structure='bigrams_to_word'
)
```

### What actually happened

```
$ cat /tmp/event.tab
cues	outcomes
#m_t#_it_mi	mit
#m_ik_k#_mi	mik
#m_n#_mi_in	min
#m_ks_mi_ik_si_i#	miksi
```

### What I expected to happen

```
$ cat /tmp/event.tab
cues	outcomes
#m_tä_it_mi_ä#	mitä
#m_ik_kä_mi_ä#	mikä
#m_ä#_mi_in_nä	minä
#m_ks_mi_ik_si_i#	miksi
```

### System details

```
Pyndl Information
=================

General Information
-------------------
Python version: 3.9.13
Pyndl version: 0.8.2

Operating System
----------------
OS: Linux x86_64
Kernel: 5.10.0-10-amd64
CPU: 8
Mem: 6461MiB/19700MiB
Swap: 0MiB/28355MiB

Dependencies
------------
numpy: 1.23.0
scipy: 1.6.1
cython: 0.29.30
pandas: 1.4.3
xarray: 2022.3.0
netCDF4: 1.6.0
pip: 21.3.1
packaging: 21.3
```

I'm a bit confused about this issue since in the JOSS paper UTF-8 is mentioned quite a bit, as well as previous studies that have successfully used the software with languages which use extended Latin. Could this be a user error somehow?
