### 2022-08-01 #4515 @frankier
I will start with a general review summary, before prioritizing the some of the GitHub issues (I have taken the liberty of including some issues not originally filed by myself here). I have expressed it in terms of what is needed for acceptance but of course there is always room for discussion.

The library looks to be quite mature overall, and have a clear potential target audience within the field of cognitive linguistics.

I think the statement of need could be a bit stronger. Some of the advantages mentioned such as maintainability are really advantages for the library author and not the library user.

A key advantage of using Python is interoperability with the Python ecosystem, however, the library does not exactly follow the conventions of any existing, opting for its own custom file-based workflow. This may be fine for the purposes of this library since it is in line with the expectation of the audience, but on the other hand much of the reusability and convenience in the Python ecosystem comes from people reusing certain idioms (e.g. that of sklearn as mentioned here https://github.com/quantling/pyndl/issues/130 ). Perhaps you could also mention the usage of/cite xarray, since this seems to be a slightly higher level package which is in use.

At a high level, I see there being 4 layers/strata in the project from the lowest:

 1. The NDL/WH rule as independent of the file format used for the events and the weights matrix (does not seem to be directly exposed/documented -- see the comments of @jinhangjiang and https://github.com/quantling/pyndl/issues/213
    * Exposing/documenting this would be a "nice to have" and could create appeal to those outside the cognitive linguistic audience
 2. The NDL/WH rule as implemented in terms of the library's custom file formats
    * This is introduced in the documentation through a word form-semantics association (meaning recognition) example
 3. Event file generation routines which can convert "END OF DOCUMENT" delimited corpora files into
    * This is introduced in the documentation between a bigram-word form association (form recognition) example
 4. Convenience functions to convert from other corpora formats (currently only OpenSubtitles) to stratum 3

I think these strata -- and therefore different ways of using the library/workflow -- should be outlined in the paper. At least the description of, and distinction between, Strata 2 and 3 should be made clear, since this is a general "what comes in"/"what comes out" description which should surely be included.

In the documentation too, the strata could be sign-posted earlier, so as to prevent confusion. Currently, Stratum 4 is only included in the API documentation, and therefore rather easy to miss. Stratum 2 and 3 are rather easy to get confused between at the moment. https://pyndl.readthedocs.io/en/latest/quickstart.html#correct-data-format deals with both. Here, the domain-purpose of the example datasets should be explained at the beginning of each example. Additionally, Strata 2 and 3 should probably be introduced briefly at the beginning of the whole before going into the examples. "Correct Data Format" => "Data Preparation". Exposing and documenting Stratum 1 may be a good longer term goal.

The possibility of calling from other languages is mentioned in the paper, however the bridge software is not mentioned by name. This should either be made clear or the sentence removed. I would favour the latter since this is really true of any Python library (in my experience it turns out to be less convenient than switching languages or using multiple languages in the repository). It would only really be worth calling attention to if you had packaged or wrapped the library specifically for these other languages. There does seem to be an example here https://pyndl.readthedocs.io/en/latest/misc.html but it looks incomplete, and people might not read anything in a section entitled "misc".

In my attempts to test the functional claims of the paper, I came across this issue https://github.com/quantling/pyndl/issues/220 . This got me thinking as to how well the package deals with Unicode in general. In particular, with things like extracting bigrams, the most reasonable thing to may most usually be to work with grapheme clusters. Note this is different from code points, which is what you will get if you work with Python 3/unicode strings directly -- if you are working with code points then you should probably switch to grapheme clusters. It might not really make sense to extract bigrams with e.g. Chinese characters as whole grapheme clusters but rather to decompose them into radicals. I'm not sure what to do about this except for document what you are doing at the moment or add some kind of caveats section to the documentation and note that multilinguality is a bit of an open box in the paper, underlining that support for ideograms / pictograms / logograms is not really there. I do think this is somewhat important given the emphasis on multilinguality in the paper.

The language of the paper is good but could still be revised a little more. If it is possible to get someone outside the work or even the field to go over it, this usually improves clarity. Here are some things I noticed:

 * Consistent capitalisation of proper nouns e.g. python/Python
 * Idiomatic expression usage e.g. "UTF-8 coded" => "UTF-8 encoded"; "developed with hindsight for usability and maintainability" => "developed with usability and maintainability in mind"
 * Usually footnotes should be preferred over brackets for asides

## Issues I believe should be addressed before publication

 1. At least one of https://github.com/quantling/pyndl/issues/211 and https://github.com/quantling/pyndl/issues/212
 2. https://github.com/quantling/pyndl/issues/208
 3. https://github.com/quantling/pyndl/issues/217
 4. https://github.com/quantling/pyndl/issues/219
 5. Putting some minimal level of benchmark in the paper (See: https://github.com/quantling/pyndl/issues/215 )
 6. Revising paper as mentioned above
 7. https://github.com/quantling/pyndl/issues/220
 8. Making sure you are working with grapheme clusters
 9. Noting lack of support for ideograms / pictograms / logograms
 10. Cite xarray

## Issues which would ideally be addressed reasonably quickly but need not necessarily block publication

 1. Putting benchmarks in the documentation (See: https://github.com/quantling/pyndl/issues/215 )
 2. https://github.com/quantling/pyndl/issues/218

## Long-term/roadmap issues

 1. https://github.com/quantling/pyndl/issues/130
 2. https://github.com/quantling/pyndl/issues/213
