### 2022-08-01 c44fcc3
Add utility function to find a non-existing filename
