### 2022-08-01 b259e69
Replace buggy file name creation by tested utility
