### 2022-08-01 quantling/pyndl#220 @derNarr
I fully agree that your expectations are completely sane and ``pyndl`` behaviour is not meeting your expectations.

Still the following code would achieve what you want to do:

```
from pyndl import preprocess

CORPUS = """
mitä
mikä
minä
miksi
""".strip()

with open("/tmp/corpus.txt", "w") as f:
    f.write(CORPUS)

if os.path.exists("/tmp/event.tab.gz"):
    os.remove("/tmp/event.tab.gz")

preprocess.create_event_file(
    corpus_file="/tmp/corpus.txt",
    event_file="/tmp/event.tab.gz",
    symbols='abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZä',
    context_structure='document',
    event_structure='consecutive_words',
    event_options=(1, ),
    cue_structure='bigrams_to_word'
)
```

For some reason we introduced a ``symbols`` parameter in ``preprocess.create_event_file`` that defaults to ascii characters only. Even if the desired behaviour can be achieved by adding the desired symbol to the ``symbols`` parameter, I will check on how to solve this so that the function defaults to allow all characters except for ``\t``, ``_``, and ``#``.
