### 2022-08-02 quantling/pyndl#220 @frankier
Okay. I see now that `symbols` is used to do a quick tokenisation of the string. Just to note that this is something that users may like to plug into and combine with external libraries (e.g. nltk, special word segmentation for CJK languages). For extracting word-like characters `\w` in regex or `isalnum()` on string are usually sufficient.
