### 2022-08-02 quantling/pyndl#220 @frankier
Also rather than specifying symbols manually users might like to e.g. provide a function which can then return true or false given a character. As an example of how this could help, it would help with filtering using data from `unicodedata` (See https://unicodebook.readthedocs.io/unicode.html#categories )
