### 2022-08-03 45f173d
Improve safe_write_path docstring
