+++ pyproject.toml
@@ -1,0 +1,66 @@
+[tool.poetry]
+name = "pyndl"
+version = "0.8.1"
+description = "Naive discriminative learning implements learning and classification models based on the Rescorla-Wagner equations."
+
+license = "MIT"
+
+authors = ["Konstantin Sering <konstantin.sering@uni-tuebingen.de>",
+           "Marc Weitz",
+           "David-Elias Künstle",
+           "Lennard Schneider",
+           "Elnaz Shafaei-Bajestan"
+           ]
+
+readme = "README.rst"  # Markdown files are supported
+
+repository = "https://github.com/quantling/pyndl"
+homepage = "https://pyndl.readthedocs.io/en/latest/"
+
+classifiers = ['Development Status :: 4 - Beta',
+               'Environment :: Console',
+               'Intended Audience :: Science/Research',
+               'Operating System :: POSIX :: Linux',
+               'Operating System :: MacOS',
+               'Topic :: Scientific/Engineering',
+               'Topic :: Scientific/Engineering :: Artificial Intelligence',
+               'Topic :: Scientific/Engineering :: Information Analysis',]
+
+build = "build.py"
+
+[tool.poetry.dependencies]
+python = ">=3.8,<3.10"  # Compatible python versions must be declared here
+numpy = "^1.23.1"
+scipy = "^1.9.0"
+pandas = "^1.4.3"
+xarray = "^2022.6.0"
+netCDF4 = "^1.6.0"
+Cython = "^0.29.32"
+
+[tool.poetry.dev-dependencies]
+pytest = "^7.0"
+pytest-cov = "^2.4"
+pydocstyle = "^6.1.1"
+flake8 = "^4.0.1"
+sphinx = "^1.4"
+sphinx_rtd_theme = "1.0.0"
+notebook = "^6.4.10"
+seaborn = "^0.11.2"
+numpydoc = "1.2"
+easydev = "0.9.35"
+pylint = "^2.0.0"
+nbsphinx = "0.8.8"
+vulture = "^2.3"
+
+
+[tool.pylint]
+  [tool.pylint.basic]
+  good-names = ["nn", "ii", "_", "jj", "df"]
+  extension-pkg-whitelist = ["numpy", "pyndl.ndl_parallel", "pyndl/ndl_openmp"]
+  ignore = ["pyndl/ndl_parallel", "pyndl/ndl_openmp"]
+
+  [tool.pylint.format]
+  max-line-length = 120
+
+  [tool.pylint.message_control]
+  disable = "E1101"
