### 2022-08-03 a3f33ce
fixes readthedocs build with pypoetry
