+++ doc/requirements.txt
@@ -1,0 +1,7 @@
+sphinx>=1.4
+sphinx_rtd_theme>=1.0.0
+notebook>=6.4.10
+seaborn>=0.11.2
+numpydoc>=1.2
+easydev>=0.9.35
+nbsphinx>=0.8.8
