+++ build.py
@@ -1,4 +1,4 @@
-from setuptools import setup, Extension
+from setuptools import Extension
 from setuptools.command.build_ext import build_ext as _build_ext
 import sys
 
