+++ tox.ini
@@ -1,58 +1,46 @@
 [tox]
+isolated_build = True
 envlist = py{38,39}-test, checkstyle, documentation
 
+
 [testenv]
-usedevelop = True
 whitelist_externals=rm
-deps =
-     test: pytest
-     -rrequirements.txt
+deps = poetry
 commands =
-     py.test --doctest-glob "*.rst"
+     poetry run pytest --doctest-glob "*.rst"
      rm doc/data/levent.tab.gz
 
 [testenv:nolinux]
-usedevelop = True
-deps =
-    pytest
-    -rrequirements.txt
 commands =
-    py.test -m "not nolinux"
+    poetry run pytest -m "not nolinux"
 
 [testenv:testdoc]
-usedevelop = True
 whitelist_externals=rm
-deps =
-    pytest
 commands =
-    py.test doc/ --doctest-glob "*.rst"
+    poetry run pytest doc/ --doctest-glob "*.rst"
     rm doc/data/levent.tab.gz
 
 [testenv:checkstyle]
-deps = pycodestyle
-commands = pycodestyle pyndl tests
+commands =
+    poetry run pycodestyle pyndl tests --ignore=E501
+ignore_outcome = True
+
 
 [testenv:travis]
-usedevelop = True
 passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
 whitelist_externals=rm
 deps =
-     mock
-     pytest
-     pytest-runner
-     sh
-     coveralls
+    poetry
+    coveralls
 commands =
-     coverage run --source=pyndl setup.py test
-     coveralls
-     py.test doc/ --doctest-glob "*.rst"
-     rm doc/data/levent.tab.gz
+    poetry run coverage run --source=pyndl -m pytest tests/
+    coveralls
+    poetry run pytest doc/ --doctest-glob "*.rst"
+    rm doc/data/levent.tab.gz
 
 [testenv:lint]
-deps =
-     pylint>=1.7.1
-     pytest
-commands = pylint [] --ignore-patterns='.*\.so' --rcfile=setup.cfg -j 2 pyndl tests
+commands =
+    poetry run pylint pyndl tests
 ignore_outcome = True
 
 [testenv:checktypes]
@@ -63,9 +51,9 @@ commands = mypy --ignore-missing-imports pyndl
 ignore_outcome = True
 
 [testenv:documentation]
-usedevelop = True
 changedir = doc
 whitelist_externals=make
-extras = docs
+deps =
+    -rdoc/requirements.txt
 commands =
     make html
