+++ pyproject.toml
@@ -64,3 +64,7 @@ vulture = "^2.3"
 
   [tool.pylint.message_control]
   disable = "E1101"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
