+++ tox.ini
@@ -2,6 +2,8 @@
 isolated_build = True
 envlist = py{38,39}-test, checkstyle, documentation
 
+[tox:.package]
+basepython = python3
 
 [testenv]
 whitelist_externals=rm
