+++ tox.ini
@@ -7,7 +7,9 @@ basepython = python3
 
 [testenv]
 whitelist_externals=rm
-deps = poetry
+deps =
+    poetry
+    pytest
 commands =
      poetry run pytest --doctest-glob "*.rst"
      rm doc/data/levent.tab.gz
@@ -33,6 +35,7 @@ passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
 whitelist_externals=rm
 deps =
     poetry
+    pytest
     coveralls
 commands =
     poetry run coverage run --source=pyndl -m pytest tests/
