+++ tox.ini
@@ -2,9 +2,6 @@
 isolated_build = True
 envlist = py{38,39}-test, checkstyle, documentation
 
-[tox:.package]
-basepython = python3
-
 [testenv]
 whitelist_externals=rm
 deps =
@@ -32,6 +29,7 @@ ignore_outcome = True
 
 [testenv:travis]
 passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
+isolated_build = True
 whitelist_externals=rm
 deps =
     poetry
