### 2022-08-03 9177396
adds explicitely installing pyndl on travis
