+++ tox.ini
@@ -3,13 +3,13 @@ isolated_build = True
 envlist = py{38,39}-test, checkstyle, documentation
 
 [testenv]
-whitelist_externals=rm
-deps =
+whitelist_externals=
+    rm
     poetry
-    pytest
 commands =
-     poetry run pytest --doctest-glob "*.rst"
-     rm doc/data/levent.tab.gz
+    poetry install -v
+    poetry run pytest --doctest-glob "*.rst"
+    rm doc/data/levent.tab.gz
 
 [testenv:nolinux]
 commands =
@@ -29,13 +29,13 @@ ignore_outcome = True
 
 [testenv:travis]
 passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
-isolated_build = True
-whitelist_externals=rm
-deps =
+whitelist_externals=
+    rm
     poetry
-    pytest
+deps = 
     coveralls
 commands =
+    poetry install -v
     poetry run coverage run --source=pyndl -m pytest tests/
     coveralls
     poetry run pytest doc/ --doctest-glob "*.rst"
