+++ .travis.yml
@@ -15,7 +15,6 @@ before_install:
   - pip3 install --upgrade pip  # all three OSes agree about 'pip3'
   - pip3 install poetry
 install:
-  - poetry install -v
   - pip3 install tox-travis
 
 script: tox -e travis
