### 2022-08-03 3448571
adds requires attribute again as it is required
