### 2022-08-03 6bedbea
adds poetry installation for non-linux build
