+++ tox.ini
@@ -12,7 +12,10 @@ commands =
     rm doc/data/levent.tab.gz
 
 [testenv:nolinux]
+whitelist_externals=
+    poetry
 commands =
+    poetry install -v
     poetry run pytest -m "not nolinux"
 
 [testenv:testdoc]
@@ -32,7 +35,7 @@ passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
 whitelist_externals=
     rm
     poetry
-deps = 
+deps =
     coveralls
 commands =
     poetry install -v
