+++ .travis.yml
@@ -1,32 +1,0 @@
-language: python
-group: travis_latest
-dist: xenial
-
-git:
-  depth: 3
-  quiet: true
-
-python:
-  - 3.9
-  - 3.8
-
-
-before_install:
-  - pip3 install --upgrade pip  # all three OSes agree about 'pip3'
-  - pip3 install poetry
-install:
-  - pip3 install tox-travis
-
-script: tox -e travis
-
-# Macos Testing
-matrix:
-  include:
-    - name: "MacOS"
-      os: osx
-      osx_image: xcode10.3  # default xcode9.4 is too old for some dependencies
-      language: sh
-      script: tox -e nolinux
-
-env:
-  OMP_NUM_THREADS: 4
