### 2022-08-03 6cb67bb
adds github actions workflow
