+++ pyproject.toml
@@ -66,5 +66,5 @@ vulture = "^2.3"
   disable = "E1101"
 
 [build-system]
-requires = ["poetry >= 0.12, <1"]
+requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.masonry.api"
