+++ pyproject.toml
@@ -67,4 +67,4 @@ vulture = "^2.3"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.masonry.api"
+build-backend = "poetry.core.masonry.api"
