+++ .github/workflows/python-test.yml
@@ -1,12 +1,6 @@
 name: tests
 
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-    branches:
-      - master
+on: [push]
 
 jobs:
   build:
