### 2022-08-03 d84b968
fixes checking distros
