### 2022-08-03 a179b0e
removes testing on Windows as currently not supported
