+++ .github/workflows/python-test.yml
@@ -7,7 +7,7 @@ jobs:
     name: ${{ matrix.os }} ${{ matrix.python-version }}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest, macOS-latest]
+        os: [ubuntu-latest, macOS-latest]
         python-version: [3.8, 3.9]
         poetry-version: [1.1.13]
       fail-fast: false
