### 2022-08-04 60faef4
adds setuptools
