+++ .travis.yml
@@ -13,7 +13,7 @@ python:
 
 before_install:
   - pip3 install --upgrade pip  # all three OSes agree about 'pip3'
-  - pip3 install poetry
+  - pip3 install poetry setuptools
 install:
   - pip3 install tox-travis
 
