### 2022-08-04 87b984a
fixes version Jinja2 due to some errors when running sphinx
