+++ pyproject.toml
@@ -51,6 +51,7 @@ easydev = "0.9.35"
 pylint = "^2.0.0"
 nbsphinx = "0.8.8"
 vulture = "^2.3"
+Jinja2 = "<3.1.0"
 
 
 [tool.pylint]
