### 2022-08-04 823aa19
cleans up sphinx config
