+++ pyproject.toml
@@ -67,5 +67,5 @@ Jinja2 = "<3.1.0"
   disable = "E1101"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["poetry>=1.0.0", "setuptools", "Cython", "numpy"]
+build-backend = "poetry.masonry.api"
