### 2022-08-05 a8d5343
updates development section about use of poetry
