### 2022-08-05 f27c592
updates release instructions
