### 2022-08-05 59efb20
Test symbol filtering in create_event_file
