### 2022-08-05 48ab313
Accept filter function for valid symbols
