### 2022-08-05 e7e67ec
removes tox as it is now replaced by poetry
