+++ tox.ini
@@ -1,65 +1,0 @@
-[tox]
-isolated_build = True
-envlist = py{38,39}-test, checkstyle, documentation
-
-[testenv]
-whitelist_externals=
-    rm
-    poetry
-commands =
-    poetry install -v
-    poetry run pytest --doctest-glob "*.rst"
-    rm doc/data/levent.tab.gz
-
-[testenv:nolinux]
-whitelist_externals=
-    poetry
-commands =
-    poetry install -v
-    poetry run pytest -m "not nolinux"
-
-[testenv:testdoc]
-whitelist_externals=rm
-commands =
-    poetry run pytest doc/ --doctest-glob "*.rst"
-    rm doc/data/levent.tab.gz
-
-[testenv:checkstyle]
-commands =
-    poetry run pycodestyle pyndl tests --ignore=E501
-ignore_outcome = True
-
-
-[testenv:travis]
-passenv = TRAVIS TRAVIS_JOB_ID TRAVIS_BRANCH
-whitelist_externals=
-    rm
-    poetry
-deps =
-    coveralls
-commands =
-    poetry install -v
-    poetry run coverage run --source=pyndl -m pytest tests/
-    coveralls
-    poetry run pytest doc/ --doctest-glob "*.rst"
-    rm doc/data/levent.tab.gz
-
-[testenv:lint]
-commands =
-    poetry run pylint pyndl tests
-ignore_outcome = True
-
-[testenv:checktypes]
-deps = mypy
-setenv =
-       MYPYPATH=./stubs/
-commands = mypy --ignore-missing-imports pyndl
-ignore_outcome = True
-
-[testenv:documentation]
-changedir = doc
-whitelist_externals=make
-deps =
-    -rdoc/requirements.txt
-commands =
-    make html
