### 2022-08-05 610ae44
Remove default symbols in event file creation
The default symbols a-zA-Z might lead to unexpected results
for users, expecting to keep unicode letter variants (e.g. accents).
