### 2022-08-05 b9924fc
Add explicit symbols to create_event_file in docs
