### 2022-08-08 51f1db6
fixes author extraction in docs build
