### 2022-08-08 51e07cf
fixes poetry development instructions in docs, adds removing created file from quickstart testing and adds pytest doctest conf options to pyproject.toml
