### 2022-08-08 1b6f163
adds comment on --runslow tests to documentation
