### 2022-08-08 a31721c
changes method=openmp to method=threading in examples to allow it to run under macos x
