+++ pyproject.toml
@@ -29,7 +29,7 @@ classifiers = ['Development Status :: 4 - Beta',
 build = "build.py"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.10"  # Compatible python versions must be declared here
+python = ">=3.8,<3.12"  # Compatible python versions must be declared here
 numpy = "^1.23.1"
 scipy = "^1.9.0"
 pandas = "^1.4.3"
