### 2022-08-08 37c4964
removes python 3.11 testing as still in beta
