+++ .github/workflows/python-test.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macOS-latest]
-        python-version: ['3.8', '3.9', '3.10', '3.11']
+        python-version: ['3.8', '3.10']
         poetry-version: [1.1.13]
       fail-fast: false
     runs-on: ${{ matrix.os }}
