### 2022-08-08 ca14941
adds info sentence about different methods
