### 2022-08-08 d0cddb2
removes double colon in develop.rst
