### 2022-08-09 b8a607d
fix runslow test of activation code
