### 2022-08-09 56d6f50
adds possibility to give a re.Pattern to symbols; makes symbols keyword-only argument; defaults symbols to be as inclusive as possible
