+++ tests/test_preprocess.py
@@ -42,24 +42,24 @@ def test_bandsample():
 def test_bad_symbols():
     with pytest.raises(ValueError):
         create_event_file(RESOURCE_FILE, EVENT_FILE,
-                          "abcd#")
+                          symbols="abcd#")
     assert not os.path.isfile(EVENT_FILE)
     with pytest.raises(ValueError):
         create_event_file(RESOURCE_FILE, EVENT_FILE,
-                          "abcd_")
+                          symbols="abcd_")
     assert not os.path.isfile(EVENT_FILE)
 
 
 def test_bad_event_context():
     with pytest.raises(NotImplementedError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE, 'a-zA-Z',
+        create_event_file(RESOURCE_FILE, EVENT_FILE, symbols='a-zA-Z',
                           context_structure="UNREASONABLE")
     assert not os.path.isfile(EVENT_FILE)
 
 
 def test_bad_event_event():
     with pytest.raises(NotImplementedError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE, 'a-zA-Z',
+        create_event_file(RESOURCE_FILE, EVENT_FILE, symbols='a-zA-Z',
                           event_structure="UNREASONABLE")
     assert not os.path.isfile(EVENT_FILE)
 
@@ -70,11 +70,13 @@ def test_drop_symbols():
     polish_event_file = os.path.join(TEST_ROOT, "temp/polish_events_corpus.tab.gz")
     symbols = "a-z"
     polish_symbols = "a-zóąćęłńśźż"
-    create_event_file(corpus_file, dropped_event_file, symbols,
+    create_event_file(corpus_file, dropped_event_file,
+                      symbols=symbols,
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
-    create_event_file(corpus_file, polish_event_file, polish_symbols,
+    create_event_file(corpus_file, polish_event_file,
+                      symbols=polish_symbols,
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -87,7 +89,8 @@ def test_drop_symbols():
 def test_symbols_filter_function():
     symbol_func = lambda chr: chr in "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
     reference_file = os.path.join(TEST_ROOT, "reference/event_file_trigrams_to_word.tab.gz")
-    create_event_file(RESOURCE_FILE, EVENT_FILE, symbol_func,
+    create_event_file(RESOURCE_FILE, EVENT_FILE,
+                      symbols=symbol_func,
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -97,7 +100,8 @@ def test_symbols_filter_function():
 
 def test_upper_case():
     event_file = os.path.join(TEST_ROOT, "temp/events_corpus_upper_case.tab.gz")
-    create_event_file(RESOURCE_FILE, event_file, 'a-zA-Z',
+    create_event_file(RESOURCE_FILE, event_file,
+                      symbols='a-zA-Z',
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -107,7 +111,8 @@ def test_upper_case():
 def test_trigrams_to_word():
     event_file = os.path.join(TEST_ROOT, "temp/event_file_trigrams_to_word.tab.gz")
     reference_file = os.path.join(TEST_ROOT, "reference/event_file_trigrams_to_word.tab.gz")
-    create_event_file(RESOURCE_FILE, event_file, 'a-zA-Z',
+    create_event_file(RESOURCE_FILE, event_file,
+                      symbols='a-zA-Z',
[…]
