### 2022-08-09 4007fc3
adds comment to clarify check
