### 2022-08-10 93383a2
adds new argument filter_func to allow for individual filter functions
