### 2022-08-11 #4515 @derNarr
@frankier thank you for your review and your good feedback. We are addressing some of your suggestions right now and might reach out to your for clarification on the others when we finished implementing the first round of improvements. Especially, on points 8. and 9. regarding the grapheme clusters and the support for ideograms / pictograms / logograms I would be happy about some more suggestions and clarifications. I will reach out to you regarding these points after we have implemented the low hanging improvements.
