### 2022-08-11 fb0f82d
renames valid_symbols into allowed_symbols
