+++ tests/test_preprocess.py
@@ -42,14 +42,14 @@ def test_bandsample():
 
 def test_bad_event_context():
     with pytest.raises(NotImplementedError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE, valid_symbols='a-zA-Z',
+        create_event_file(RESOURCE_FILE, EVENT_FILE, allowed_symbols='a-zA-Z',
                           context_structure="UNREASONABLE")
     assert not os.path.isfile(EVENT_FILE)
 
 
 def test_bad_event_event():
     with pytest.raises(NotImplementedError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE, valid_symbols='a-zA-Z',
+        create_event_file(RESOURCE_FILE, EVENT_FILE, allowed_symbols='a-zA-Z',
                           event_structure="UNREASONABLE")
     assert not os.path.isfile(EVENT_FILE)
 
@@ -61,12 +61,12 @@ def test_drop_symbols():
     symbols = "a-z"
     polish_symbols = "a-zóąćęłńśźż"
     create_event_file(corpus_file, dropped_event_file,
-                      valid_symbols=symbols,
+                      allowed_symbols=symbols,
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
     create_event_file(corpus_file, polish_event_file,
-                      valid_symbols=polish_symbols,
+                      allowed_symbols=polish_symbols,
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -80,7 +80,7 @@ def test_symbols_filter_function():
     symbol_func = lambda chr: chr in "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
     reference_file = os.path.join(TEST_ROOT, "reference/event_file_trigrams_to_word.tab.gz")
     create_event_file(RESOURCE_FILE, EVENT_FILE,
-                      filter_func=symbol_func,
+                      allowed_symbols=symbol_func,
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -91,7 +91,7 @@ def test_symbols_filter_function():
 def test_upper_case():
     event_file = os.path.join(TEST_ROOT, "temp/events_corpus_upper_case.tab.gz")
     create_event_file(RESOURCE_FILE, event_file,
-                      valid_symbols='a-zA-Z',
+                      allowed_symbols='a-zA-Z',
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -102,7 +102,7 @@ def test_trigrams_to_word():
     event_file = os.path.join(TEST_ROOT, "temp/event_file_trigrams_to_word.tab.gz")
     reference_file = os.path.join(TEST_ROOT, "reference/event_file_trigrams_to_word.tab.gz")
     create_event_file(RESOURCE_FILE, event_file,
-                      valid_symbols='a-zA-Z',
+                      allowed_symbols='a-zA-Z',
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ),
@@ -115,7 +115,7 @@ def test_trigrams_to_word_line_based():
     event_file = os.path.join(TEST_ROOT, "temp/event_file_trigrams_to_word_line_based.tab.gz")
     reference_file = os.path.join(TEST_ROOT, "reference/event_file_trigrams_to_word_line_based.tab.gz")
     create_event_file(RESOURCE_FILE, event_file,
-                      valid_symbols='a-zA-Z',
+                      allowed_symbols='a-zA-Z',
                       context_structure="document",
                       event_structure="line", event_options=(3, ),
                       cue_structure="trigrams_to_word")
@@ -127,7 +127,7 @@ def test_bigrams_to_word():
[…]
