+++ doc/source/examples.rst
@@ -276,7 +276,7 @@ generate an event file based on a raw corpus file and filter it:
     >>> from pyndl import preprocess
     >>> preprocess.create_event_file(corpus_file='doc/data/lcorpus.txt',
     ...                              event_file='doc/data/levent.tab.gz',
-    ...                              valid_symbols='a-zA-Z',
+    ...                              allowed_symbols='a-zA-Z',
     ...                              context_structure='document',
     ...                              event_structure='consecutive_words',
     ...                              event_options=(1, ),
