### 2022-08-11 dc5acf7
Fix #220 create_event_file removes non-ascii letters by default (#224)
- We rename `symbols` to `allowed_symbols` to make this argument more self-explainable
- We do not introduce a DeprecationWarning as the goal is to release v1.0.0 as soon as possible anyway and no other version will be released inbetween
- `allowed_symbols` will be a keyword-only argument as being inclusive here is the default to expect
- `filter_func` is now also part of `allowed_symbols` aka `allowed_symbols` can also be a function that indicates whether a character is allowed or not
- test coverage is increased and expli
[…]
