+++ tests/test_preprocess.py
@@ -5,6 +5,7 @@
 from collections import OrderedDict
 import gzip
 import os
+import re
 
 import pytest
 
@@ -39,34 +40,58 @@ def test_bandsample():
     bandsample(outcome_freq_map, 50, cutoff=1, verbose=True)
 
 
-def test_bad_symbols():
-    with pytest.raises(ValueError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE,
-                          "abcd#")
-    assert not os.path.isfile(EVENT_FILE)
-    with pytest.raises(ValueError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE,
-                          "abcd_")
-    assert not os.path.isfile(EVENT_FILE)
-
-
 def test_bad_event_context():
     with pytest.raises(NotImplementedError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE,
+        create_event_file(RESOURCE_FILE, EVENT_FILE, allowed_symbols='a-zA-Z',
                           context_structure="UNREASONABLE")
     assert not os.path.isfile(EVENT_FILE)
 
 
 def test_bad_event_event():
     with pytest.raises(NotImplementedError):
-        create_event_file(RESOURCE_FILE, EVENT_FILE,
+        create_event_file(RESOURCE_FILE, EVENT_FILE, allowed_symbols='a-zA-Z',
                           event_structure="UNREASONABLE")
     assert not os.path.isfile(EVENT_FILE)
+
+
+def test_drop_symbols():
+    corpus_file = RESOURCE_FILE  # includes polish symbols
+    dropped_event_file = os.path.join(TEST_ROOT, "temp/dropped_events_corpus.tab.gz")
+    polish_event_file = os.path.join(TEST_ROOT, "temp/polish_events_corpus.tab.gz")
+    symbols = "a-z"
+    polish_symbols = "a-zóąćęłńśźż"
+    create_event_file(corpus_file, dropped_event_file,
+                      allowed_symbols=symbols,
+                      context_structure="document",
+                      event_structure="consecutive_words",
+                      event_options=(3, ))
+    create_event_file(corpus_file, polish_event_file,
+                      allowed_symbols=polish_symbols,
+                      context_structure="document",
+                      event_structure="consecutive_words",
+                      event_options=(3, ))
+    with pytest.raises(AssertionError):  # polish file should contain different events
+        compare_event_files(dropped_event_file, polish_event_file)
+    os.remove(dropped_event_file)
+    os.remove(polish_event_file)
+
+
+def test_symbols_filter_function():
+    symbol_func = lambda chr: chr in "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+    reference_file = os.path.join(TEST_ROOT, "reference/event_file_trigrams_to_word.tab.gz")
+    create_event_file(RESOURCE_FILE, EVENT_FILE,
+                      allowed_symbols=symbol_func,
+                      context_structure="document",
+                      event_structure="consecutive_words",
+                      event_options=(3, ))
+    compare_event_files(EVENT_FILE, reference_file)
+    os.remove(EVENT_FILE)
 
 
 def test_upper_case():
     event_file = os.path.join(TEST_ROOT, "temp/events_corpus_upper_case.tab.gz")
     create_event_file(RESOURCE_FILE, event_file,
+                      allowed_symbols='a-zA-Z',
                       context_structure="document",
                       event_structure="consecutive_words",
                       event_options=(3, ))
@@ -77,6 +102,7 @@ def test_trigrams_to_word():
[…]
