### 2022-08-11 9498844
renames test workflow and restricts runs to PRs and master branch
