+++ .github/workflows/python-test.yml
@@ -1,6 +1,12 @@
-name: tests
+name: Linux & OSX tests
 
-on: [push]
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
 
 jobs:
   build:
