### 2022-08-11 1fa521b
bumps version to 1.0.0-beta
