+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyndl"
-version = "0.8.3"
+version = "1.0.0-beta"
 description = "Naive discriminative learning implements learning and classification models based on the Rescorla-Wagner equations."
 
 license = "MIT"
