### 2022-08-11 610d906
fixes authors in docs
