### 2022-08-11 a47618c
Adds support for pypoetry (https://python-poetry.org) (#222)
* introduces python poetry to maintain and build the package

* allows building wheels for PyPI  (fixes #212) 

* uses `pyproject.toml` instead of `setup.py` and thus fixes  #143 and fixes #137 

* sets up Github Actions for general testing (replaces Travis)

* bumps version to 1.0.0-beta

Co-authored-by: Konstantin (Tino) Sering <konstantin.sering@uni-tuebingen.de>
