+++ doc/source/development.rst
@@ -1,10 +1,10 @@
 Development
 ===========
-.. image:: https://travis-ci.com/quantling/pyndl.svg?branch=master
-    :target: https://travis-ci.com/quantling/pyndl?branch=master
+.. image:: https://github.com/quantling/pyndl/actions/workflows/python-test.yml/badge.svg?branch=master
+    :target: https://github.com/quantling/pyndl/actions/workflows/python-test.yml
 
-.. image:: https://coveralls.io/repos/github/quantling/pyndl/badge.svg?branch=master
-    :target: https://coveralls.io/github/quantling/pyndl?branch=master
+.. image:: https://codecov.io/gh/quantling/pyndl/branch/master/graph/badge.svg?token=2GWUXRA9PD
+    :target: https://codecov.io/gh/quantling/pyndl
 
 .. image:: https://img.shields.io/lgtm/grade/python/g/quantling/pyndl.svg?logo=lgtm&logoWidth=18
     :target: https://lgtm.com/projects/g/quantling/pyndl/context:python
@@ -39,14 +39,13 @@ Workflow
 1. Fork this repository on Github. From here on we assume you successfully
    forked this repository to https://github.com/yourname/pyndl.git
 
-2. Get a local copy of your fork and install the package in 'development'
-   mode, which will make changes in the source code active immediately, by running
+2. Install all dependencies with poetry (https://python-poetry.org/)
 
    .. code:: bash
 
        git clone https://github.com/yourname/pyndl.git
        cd pyndl
-       python setup.py develop --user
+       poetry install
 
 3. Add code, tests or documentation.
 
@@ -54,8 +53,8 @@ Workflow
 
    .. code:: bash
 
-       make checkstyle
-       make test
+        poetry run pytest
+        poetry run pylint pyndl
 
 5. Add and commit your changes after tests run through without complaints.
 
@@ -78,46 +77,44 @@ Workflow
    click "New pull request" to submit your Pull Request to
    https://github.com/quantling/pyndl.
 
-.. note::
-
-    If you want to develop *pyndl* you should have installed:
-
-    .. code:: bash
-
-        pip install --user tox pylint pytest pycodestyle sphinx
-
 
 Running tests
 -------------
 
-We use ``make`` and ``tox`` to manage testing. You can run the tests by
+We use ``poetry`` to manage testing. You can run the tests by
 executing the following within the repository's root folder (``pyndl/``):
 
 .. code:: bash
 
-    make test
+    poetry run pytest
+
+For extensive, time and memory consuming tests run (at least 12 GB of free
+memory should be available):
+
+.. code:: bash
+
+    poetry run pytest --run-slow
 
 For manually checking coding guidelines run:
 
 .. code:: bash
 
-    make checkstyle
-
-There is an additional way to invoke ``pylint`` as a linter with tox by running
-
-.. code:: bash
-
-    tox -e lint
+    poetry run pylint pyndl
 
 The linting gives still a lot of complaints that need some decisions on how to
 fix them appropriately.
+
+.. note::
+
+    Previous versions of *pyndl* used ``make`` and ``tox`` to manage testing. For
+    documentation on this, please check the respective version documentations
 
 
 Local testing with conda
 ------------------------
 
 Sometimes it might be useful to test if ``pyndl`` works in a clean python
-environment. Besides ``tox`` this is possible with ``conda`` as well. The
+environment. Besides ``poetry`` this is possible with ``conda`` as well. The
 commands are as follows:
 
 .. code:: bash
@@ -201,26 +198,27 @@ Some sources with more explanation:
 Building documentation
 ----------------------
 
-Building the documentation requires some extra dependencies. Therefore, run
+Building the documentation requires some extra dependencies. Usually, these are
[…]
