+++ setup.cfg
@@ -1,13 +1,0 @@
-[pycodestyle]
-ignore =
-max-line-length = 120
-
-[pylint]
-max-line-length = 120
-good-names = nn, ii, _, jj, df
-extension-pkg-whitelist=numpy,pyndl.ndl_parallel,pyndl/ndl_openmp
-ignore=pyndl/ndl_parallel,pyndl/ndl_openmp
-disable=E1101
-
-[aliases]
-test=pytest

[…]
