### 2022-08-11 d89514c
bumps version to 1.0.0
