### 2022-08-11 4ac49b8
removes deprecated arguments and warnings
