+++ pyndl/activation.py
@@ -18,7 +18,7 @@ from . import io
 
 
 # pylint: disable=W0621
-def activation(events, weights, *, n_jobs=1, number_of_threads=None,
+def activation(events, weights, *, n_jobs=1,
                remove_duplicates=None, ignore_missing_cues=False):
     """
     Estimate activations for given events in event file and outcome-cue weights.
@@ -60,11 +60,6 @@ def activation(events, weights, *, n_jobs=1, number_of_threads=None,
         returned if weights is instance of dict
 
     """
-    if number_of_threads is not None:
-        warnings.warn("Parameter `number_of_threads` is renamed to `n_jobs`. The old name "
-                      "will stop working with v0.9.0.",
-                      DeprecationWarning, stacklevel=2)
-        n_jobs = number_of_threads
     if isinstance(events, str):
         events = io.events_from_file(events)
 
