### 2022-08-11 quantling/pyndl#220 @derNarr
Hopefully, our changes fix this issue and prevent other new users to run into the same problem.

For more complex tokenization tasks it is strongly advised to create an event file by hand. In the end, the event file is a tab separated two column table file, where in each column different cues or outcomes are separated by an underscore. Depending on the source files this is two `'_'.join(...)` and one `data.to_csv('path_to_event_file.tab.gz', sep='\t')` line of code.

But feel free to reopen this issue, if there are still concerns left.
