### 2022-08-15 #4515 @VenkteshV
I am finishing up my reviews. I have some recommendations for authors (minor) on the writing part:

1) You could consider making the statement of need a bit more detailed. The summary stresses the non-english support component. However, I find this missing in the statement of need. More emphasis can also be made on size of corpora the algorithm can handle.
2) You could also consider splitting the second paragraph into listing existing software and stating the clear differences between pyndl and others.
@derNarr
