### 2022-08-21 quantling/pyndl#226 @VenkteshV
Instructions to run tests can be mentioned in the readme.txt of test folder
This is a part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4515).  The code is well written. But in the tests folder i see a Readme.txt containing an itenary of the documents in that folder. It would also be useful if the command to run the tests can be specified in this readme. Though i agree it is in the documentation site it would be a nice to have in the repo.
