### 2022-08-29 #4515 @derNarr
@osorensen , I am handing in writing my PhD thesis and will submit it in mid September. Therefore, I will addressing the issues of the reviewers in the end of September / beginning of October. I missed that @VenkteshV 's evaluation of the code is in issue https://github.com/quantling/pyndl/issues/226 and was still waiting for it.

Thanks to all of you for the good suggestions and the reviews. We already started addressing some of the issues raised and I will write some prose here, when we think that we have made progress in addressing some of these issues.
