### 2022-08-29 quantling/pyndl#227 @dekuenstle
Revise JOSS paper
During the JOSS review, we received several recommendations for the paper (esp. [here](https://github.com/openjournals/joss-reviews/issues/4515#issuecomment-1201165829)):

- [ ] mention interop with python ecosystem (e.g. cite xarray)
- [ ] outline the different workflows to use pyndl (see @frankier's comment). Check/add to the docs, too!
- [ ] Either remove or specify sentence about calling other languages. Move example in docs from MISC to a better place.
- [ ]"Consistent capitalisation of proper nouns e.g. python/Python"
- [ ] "Idiomatic expression usage e.g. "UTF-8 coded" => "UTF-8 encoded"; "developed with hindsight for usability and maintainability" => "developed with usability and maintainability in mind"
- [ ] "Usually footnotes should be preferred over brackets for asides"
- [ ] "Noting lack of support for ideograms / pictograms / logograms". Also in the docs
