### 2022-08-29 quantling/pyndl#228 @dekuenstle
Explain usage of grapheme clusters with non-western languages
@frankier raised questions about how to use grapheme clusters, e.g. with Chinese characters, in pyndl.

We should stress in the documentation and the JOSS paper, that in this case the Chinese characters should already be preprocessed (by any other software) ––pyndl expects to deal with grapheme-like content (e.g. radicals in this case).

See [frankier's description](https://github.com/openjournals/joss-reviews/issues/4515):

> In my attempts to test the functional claims of the paper, I came across this issue https://github.com/quantling/pyndl/issues/220 . This got me thinking as to how well the package deals with Unicode in general. In particular, with things like extracting bigrams, the most reasonable thing to may most usually be to work with grapheme clusters. Note this is different from code points, which is what you will get if you work with Python 3/unicode strings directly -- if you are working with code points then you should probably switch to grapheme clusters. It might not really make sense to extract bigrams with e.g. Chinese characters as whole grapheme clusters but rather to decompose them into radicals. I'm not sure what to do about this except for document what you are doing at the moment or add some kind of caveats section to the documentation and note that multilinguality is a bit of an open box in the paper, underlining that support for ideograms / pictograms / logograms is not really there. I do think this is somewhat important given the emphasis on multilinguality in the paper.
