### 2022-11-10 quantling/pyndl#228 @derNarr
The main underlying misunderstanding seems to be that a major focus of `pyndl` is on the clever creation of cues and outcomes for different languages that can be learned with naive disciminative learning. From our point of view the main focus is on implementing the learning itself in an efficient way and the preprocessing of the best inputs (cues) and outputs (outcomes) should be done with another software.
