### 2022-11-10 9094545
addresses issue #227 and improves the paper
