+++ paper/paper.bib
@@ -1,3 +1,17 @@
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {
+https://doi.org/10.5334/jors.148
+}
+}
+
 @incollection{rescorla1972theory,
   title={A theory of Pavlovian conditioning: Variations in the effectiveness of reinforcement and nonreinforcement},
   author={Rescorla, Robert A and Wagner, Allan R},
@@ -161,6 +175,23 @@
 	author = {R. Harald Baayen and Michael Ramscar},
 	title = {5. Abstraction, storage and naive discriminative learning},
 	booktitle = {Handbook of Cognitive Linguistics}
+}
+
+@article{Arnold_2017,
+	doi = {10.1371/journal.pone.0174623},
+	url = {
+https://doi.org/10.1371%2Fjournal.pone.0174623
+},
+	year = 2017,
+	month = {apr},
+	publisher = {Public Library of Science ({PLoS})},
+	volume = {12},
+	number = {4},
+	pages = {e0174623},
+	author = {Denis Arnold and Fabian Tomaschek and Konstantin Sering and Florence Lopez and R. Harald Baayen},
+	editor = {Hedderik van Rijn},
+	title = {Words from spontaneous conversational speech can be recognized with human-like accuracy by an error-driven learning algorithm that discriminates between meanings straight from smart acoustic features, bypassing the phoneme as recognition unit},
+	journal = {{PLOS} {ONE}}
 }
 
 @article{Rescorla1988PavlovianCI,
