### 2022-11-10 6817151
Add CodeQL workflow for GitHub code scanning
