### 2022-11-14 0066f06
updates documentation section names and reorders sections to a clearer order (fixes #217)
