### 2022-11-14 quantling/pyndl#227 @derNarr
- [ ] inconsistent use of American English and British English (we should use American English)
