### 2022-11-14 604bb41
first round of improving the WH documentation
