+++ pyndl/wh.py
@@ -4,8 +4,10 @@ pyndl.wh
 
 *pyndl.wh* provides functions in order to train Widrow-Hoff (WH) models. In contrast
 to the Rescorla-Wagner (RW) models in the WH models can not only have binary
-cues and outcomes, but can code gradual intesities in there cues and outcomes
-as real, continues values.
+cues and outcomes, but can encode gradual intesities in the cues and outcomes.
+This is done by associating a vector of continues values (real numbers) to each
+cue and outcome. The size of the vector has to be the same for all cues and for
+all outcomes, but can differ between cues and outcomes.
 
 It is possible to calculate weights for continuous cues or continues outcomes,
 while keeping the outcomes respectively cues binary. Finally, it is possible to
@@ -54,15 +56,15 @@ def wh(events, eta, *, cue_vectors=None, outcome_vectors=None,
         events_per_temporary_file=10000000):
     """
     Calculate the weights for all events using the Widrow-Hoff learning rule in
-    three different flavours.
+    three different flavors.
 
-    In the first flavour, cues and outcomes both are vectors and the names in
+    In the first flavor, cues and outcomes both are vectors and the names in
     the eventfiles refer to these vectors. The vectors for all cues and
-    outcomes are given as an xarrayDataArray with the arguments `cue_vectors`
+    outcomes are given as an xarray.DataArray with the arguments `cue_vectors`
     and `outcome_vectors'.
 
-    In the second and third flavour, only the cues or only the outcomes are
-    treated as vectors but the ones not being treated as vectors are still
+    In the second and third flavor, only the cues or only the outcomes are
+    treated as vectors and the ones not being treated as vectors are still
     considered being present or not being present in a binary way.
 
     This is a parallel python implementation using cython, numpy,
@@ -81,23 +83,24 @@ def wh(events, eta, *, cue_vectors=None, outcome_vectors=None,
     method : {'openmp', 'threading', 'numpy'}
         'numpy' works only for real to real Widrow-Hoff.
     weights : None or xarray.DataArray
-        the xarray.DataArray needs to have the dimensions 'cues' and 'outcomes'
+        the xarray.DataArray needs to have the named dimensions 'cues' or
+        'cue_vector_dimensions' and 'outcomes' or 'outcome_vector_dimensions'
     n_jobs : int
         an integer giving the number of threads in which the job should be
         executed
     n_outcomes_per_job : int
         an integer giving the number of outcomes that are processed in one job
     remove_duplicates : {None, True, False}
-        if None though a ValueError when the same cue is present multiple times
+        if None raise a ValueError when the same cue is present multiple times
         in the same event; True make cues and outcomes unique per event; False
         keep multiple instances of the same cue or outcome (this is usually not
         preferred!)
     verbose : bool
-        print some output if True.
+        print some output if True
     temporary_directory : str
         path to directory to use for storing temporary files created;
         if none is provided, the operating system's default will
-        be used (/tmp on unix)
+        be used like '/tmp' on unix
     events_per_temporary_file: int
         Number of events in each temporary binary file. Has to be larger than 1
 
@@ -172,8 +175,8 @@ def dict_wh(events, eta, cue_vectors, outcome_vectors, *,
     a fast cython implementation.
 
     The main purpose of this function is to have a reference implementation
-    which is used to test the faster cython version against. Additionally, this
-    function can be a good starting point to develop different flavours of the
+    which is used to validate the faster cython version against. Additionally, this
[…]
