### 2022-11-15 770361b
adds WH example; improves docstrings in Widrow-Hoff code (wh.py); checks dtype of cue_vectors and outcome_vectors
