### 2022-11-16 quantling/pyndl#228 @frankier
I agree that it is reasonable to decide that text processing and cue creation is out of scope of the library. However, if you take the position that the library is really only about learning, then the paper need not put any particular emphasis on UTF-8 support, since this primes people to think that good support of different writing systems is a major aim of the library. There might be some in-between position you can take here too.
