### 2022-11-16 8076b0e
Feat: Option for Frequency col in event files
ndl and ndl2 require this column. This extension increases
interoperabilty between NDL implementations
