### 2022-11-16 ba2d262
Add benchmark scripts with results on lenovo laptop
