### 2022-11-16 92d803f
Add benchmark description to docs
