### 2022-11-16 3b58c9e
Add benchmark to paper
