+++ paper/paper.md
@@ -78,8 +78,12 @@ researchers.
 However, the code of *ndl2* is only available upon request. One reason for this
 has been that it only runs on Linux and CRAN's guidelines make it difficult to
 publish single platform packages [@R_project]. Another reason is the limited
-maintainability through low level C and C++ code next to high level R code.  A
-severe limitation of *ndl* and *ndl2* is that both packages have difficulties
+maintainability through low level C and C++ code next to high level R code. 
+The better maintainability of *pyndl* does not come at the cost of performance: 
+The benchmark results in \autoref{fig:benchmark}, described in detail in our package's documentation, shows that *pyndl* is faster than *ndl* and *ndl2*. 
+Memory is used efficiently by storing data records in compressed form and loading data points as they are needed during learning.
+
+A severe limitation of *ndl* and *ndl2* is that both packages have difficulties
 with non-ASCII input, causing problems in the analysis of non-English text
 corpora due to special characters or non-Latin alphabets. An example would be
 the processing of Arabic or Ukrainian languages; even German umlauts are
@@ -89,6 +93,7 @@ of arrays in the R programming language [@R_project]. This limit does not allow
 for more than 46,340 word types in a word-type to word-type model, which is too
 small to capture the full lexicon in most languages.
 
+![Runtime mean and standard-error ($n$=10) for learning trigram-to-word models with different NDL implementations. Ours, *pyndl*, is the fastest for large datasets.\label{fig:benchmark}](benchmark_result.png)
 
 # Implementation and use in research
 
