### 2022-11-16 b4baa02
Prettier plot
