### 2022-11-16 a5ba3a1
fixes #226; improves tests/README.rst and explains how to run tests here and give a reference to the documentation
