### 2022-11-16 quantling/pyndl#211 @derNarr
As far as I understand this, it boils down to the ability to properly compile Python extensions with Cython. In order to be able to properly compile these extensions Cython, Python headers and a C compiler like gcc or clang need to be installed. If you use a scientific Python distribution like Miniconda this distribution seems to taken care of the necessary steps, e.g. running `conda install cython` should give you all you need to properly compile extensions.

With the next update of `pyndl` this will be pointed out more clearly in the documentation. Furthermore, we added wheels (for Linux), which does not need the compile step at all.

This issue will be closed, when the pull request is merged that adds the information to the documentation. @frankier if you feel I missed an important aspect, please reopen the issue and point out what is missing. Thank you! :-)
