### 2022-11-16 quantling/pyndl#228 @derNarr
Besides that it is easier to maintain and explore different variants of the RW learning rule in a Python/Cython library compared to a R/Rcpp/C++ library one of the major drawbacks of the `ndl2` library is its poor support for Unicode. Therefore, one of my main selling point to people, who used `ndl2` in their research is/was the improved Unicode support in the `pyndl` package.

But I see now, that this emphasis only makes sense in the context of having the poor Unicode support in R and `ndl2` but gives the wrong suggestions to someone used to Python3, where Unicode strings became a language feature and therefore is not really worth mentioning anymore.

I am still a bit puzzled on how to properly address these two audiences. The one audience coming from Python3, who are rightfully underwhelmed that `pyndl` can use Unicode and in contrast the audience coming from R and `ndl2` for whom this might be a good reason to switch to a different programming language.

I will try to make this in-between position more clear and transparent. @frankier if you have any suggestions, I am happy to hear about them as well :-)
