### 2022-11-16 a95d26c
fixes #187 and fixes #211; adds notes on the importants of proper Cython installations as a dependency of pyndl; removes the suggestion of install gcc on MacOS as clang should be fine now
