### 2022-11-16 47bb6ac
moves the benchmark to "Implementation and use in research" and adds a longer caption to the figure
