### 2022-11-17 6740e09
restructures quickstart.rst and adds overview paragraph
