### 2022-11-17 cd5aa3f
Improves WH documenation (#232)
* adds WH example; improves docstrings in Widrow-Hoff code (wh.py); checks dtype of cue_vectors and outcome_vectors

* fixes #226; improves tests/README.rst and explains how to run tests here and give a reference to the documentation

* fixes #187 and fixes #211; adds notes on the importance of proper Cython installations as a dependency of pyndl; removes the suggestion to install gcc on MacOS as clang should be fine now

Co-authored-by: Marc Weitz <marc.weitz@onlinehome.de>
