### 2022-11-17 b517c40
Improves documentation (#230)
* updates documentation section names and reorders sections to a clearer order (fixes #217)

* clarifies quickstart section

* restructures quickstart.rst and adds overview paragraph

* adds summary from paper to package doc string

* moves quickstart description to start page
