+++ doc/source/index.rst
@@ -1,15 +1,17 @@
 .. automodule:: pyndl
   :noindex:
 
+.. include:: ./quickstart.rst
+
 .. toctree::
    :maxdepth: 2
+   :hidden:
    :caption: Contents:
 
-   quickstart
    installation
    ndl
-   examples
+   examples 
+   tips_and_tricks
    pyndl
    development
-   misc
-   about
+   credits
