### 2022-11-17 e778154
Benchmark  (#233)
* Add benchmark scripts with results on lenovo laptop
* Add benchmark description to docs
* Add plots to docs and paper

Co-authored-by: Konstantin (Tino) Sering <konstantin.sering@uni-tuebingen.de>
