+++ pyndl/ndl.py
@@ -147,7 +147,7 @@ def ndl(events, alpha, betas, lambda_=1.0, *,
 
     if not (remove_duplicates is None or isinstance(remove_duplicates, bool)):
         raise ValueError("remove_duplicates must be None, True or False")
-    if not isinstance(events, str):
+    if not isinstance(events, (str, os.PathLike)):
         raise ValueError("'events' need to be the path to a gzipped event file not {}".format(type(events)))
 
     weights_ini = weights
