+++ doc/source/benchmark.rst
@@ -1,0 +1,145 @@
+Benchmark
+=========
+
+Here we compare the performance of Naive Discrimination Learning in `pyndl`
+with other implementations, namely from the `R` packages `ndl` and `ndl2`. 
+
+This document summarizes the benchmarking procedure and results; the code to run
+these is available in the repository's `benchmark folder <https://github.com/quantling/pyndl/tree/master/benchmark>`_.  
+
+
+Creating the event files
+------------------------
+
+The benchmark datasets are trigram-to-word events on the text of the 
+German *Graphem* Wikipedia page, that can be generated with ``python preprocessing.py``.
+
+::
+    
+    Cues: ['aph', 'em#', 'hem', 'gra', 'rap', '#gr', 'phe'], Outcome: ['graphem']
+    Cues: ['aus', '#au', 'us#'], Outcome: ['aus']
+    Cues: ['wik', '#wi', 'kip', 'iki', 'edi', 'ia#', 'ped', 'ipe', 'dia'], Outcome: ['wikipedia']
+    …
+    Cues: ['ქარ', '#ქა', 'ართ', 'ლი#', 'რთუ', 'თულ', 'ული'], Outcome: ['ქართული']
+    Cues: ['ша#', 'қша', 'ақш', '#қа', 'зақ', 'қаз', 'аза'], Outcome: ['қазақша']
+    Cues: ['#한국', '한국어', '국어#'], Outcome: ['한국어']
+    …
+
+The text contains some non-ASCII characters from various languages. 
+While this is not a problem with `pyndl`, these characters would crash `ndl2`.
+This is why create cue-outcome events with a basic preprocessing, 
+including just ASCII characters and letters.
+Please note, that in research applications we recommend using dedicated software for more careful
+preprocessing.
+
+.. code-block:: python
+    
+    preprocess.create_event_file(corpus_file=txt_file,
+                                 event_file=ascii_event_file,
+                                 allowed_symbols='a-zA-Z0-9',
+                                 context_structure='document',
+                                 event_structure='consecutive_words',
+                                 event_options=(1,),  # number of words,
+                                 cue_structure="trigrams_to_word",
+                                 lower_case=True,
+                                 remove_duplicates=True)
+
+Based on this event file, 
+we construct large scale event file by concatenating all events multiple times.
+
+
+.. code-block:: python
+
+    ascii_events = list(io.events_from_file(ascii_event_file))
+    io.events_to_file(ascii_events * times, event_dir / f"{times}_times_{ascii_event_file.name}", compatible=True)
+    
+    
+Running pyndl
+-------------
+
+Our benchmark compares the runtime of `pyndl`'s NDL implementations on event files of varying size. 
+The naive `dict_ndl` learner is orders of magnitude slower and thus excluded from the comparison.
+In addition to single processing,  both parallel processing methods
+(*threading* and *OpenMP*) are included. Run the benchmark with ``python run_ndl.py``, which
+saves the wall-clock times in a CSV files.
+
+
+.. code-block:: python
+
+    def clock(func, args, **kwargs):
+        gc.collect()
+        start = time.time()
+        result = func(*args, **kwargs)
+        stop = time.time()
+        duration = stop - start
+        return result, duration
+    …
+    for r in range(repeats):
+        for file_path in event_dir.glob('*.tab.gz'):
+            _, duration_omp1 = clock(ndl.ndl, (file_path, ALPHA, BETAS, LAMBDA_), n_jobs=1, method='openmp') 
+            _, duration_omp4 = clock(ndl.ndl, (file_path, ALPHA, BETAS, LAMBDA_), n_jobs=4, method='openmp') 
+            …
+
[…]
