### 2022-11-18 6aba4b3
adds small improvements of the quickstart section (#235)
