### 2022-11-18 quantling/pyndl#219 @Trybnetic
We hope that we addressed this issue with our [updated quickstart section](https://pyndl.readthedocs.io/en/latest/) where we aim to outline the purpose of the examples and the different steps/strata now in more detail.

You are absolutely right that there is on one hand the Naive Discriminative Learning (NDL) rule which is the main focus of this package. As over the years some preprocessing methods got established and are now often used, at least in our group, these were also included in this package, but are not required to run the NDL rule. However, our implementation of NDL expects the data to be in a certain format. How this format is obtained (through one of the provided established preprocessing methods or by a custom one) is up to the researcher and even more importantly up to the research question. We hope this becomes now clearer when reading the quickstart section.

We would very happy about your feedback whether this resolves your issue or whether we missed something here.
