### 2022-11-22 6acb964
renames doc folder to docs (#234)
