+++ docs/source/conf.py
@@ -68,5 +68,5 @@ html_context = {"display_github": True,  # Add 'Edit on Github' link instead of 
                 "github_user": 'quantling',
                 "github_repo": 'pyndl',
                 "github_version": 'master',
-                "conf_py_path": '/doc/source/'
+                "conf_py_path": '/docs/source/'
                 }
