+++ docs/source/examples.rst
@@ -123,7 +123,7 @@ Let's start:
 .. code-block:: python
 
     >>> from pyndl import ndl
-    >>> weights = ndl.ndl(events='doc/data/lexample.tab.gz', alpha=0.1,
+    >>> weights = ndl.ndl(events='docs/data/lexample.tab.gz', alpha=0.1,
     ...                   betas=(0.1, 0.1), method='threading')
     >>> weights  # doctest: +ELLIPSIS
     <xarray.DataArray (outcomes: 8, cues: 15)>
@@ -165,7 +165,7 @@ weight matrix by specifying the ``weight`` argument:
 
 .. code-block:: python
 
-    >>> weights2 = ndl.ndl(events='doc/data/lexample.tab.gz', alpha=0.1,
+    >>> weights2 = ndl.ndl(events='docs/data/lexample.tab.gz', alpha=0.1,
     ...                    betas=(0.1, 0.1), method='threading', weights=weights)
     >>> weights2  # doctest: +ELLIPSIS
     <xarray.DataArray (outcomes: 8, cues: 15)>
@@ -208,7 +208,7 @@ Therefore
 
 .. code-block:: python
 
-    >>> weights = ndl.dict_ndl(events='doc/data/lexample.tab.gz',
+    >>> weights = ndl.dict_ndl(events='docs/data/lexample.tab.gz',
     ...                        alphas=0.1, betas=(0.1, 0.1))
     >>> weights['plural']['s#'] # doctes: +ELLIPSIS
     0.076988227...
@@ -221,9 +221,9 @@ learning or do both:
 
     >>> alphas_cues = dict(zip(['#h', 'ha', 'an', 'nd', 'ds', 's#', '#l', 'la', 'as', 'ss', 'ad', 'd#', '#a', '#s', 'sa'],
     ...                        [0.1, 0.2, 0.3, 0.4, 0.1, 0.2, 0.3, 0.1, 0.2, 0.1, 0.2, 0.1, 0.3, 0.1, 0.2]))
-    >>> weights = ndl.dict_ndl(events='doc/data/lexample.tab.gz',
+    >>> weights = ndl.dict_ndl(events='docs/data/lexample.tab.gz',
     ...                        alphas=alphas_cues, betas=(0.1, 0.1))
-    >>> weights2 = ndl.dict_ndl(events='doc/data/lexample.tab.gz',
+    >>> weights2 = ndl.dict_ndl(events='docs/data/lexample.tab.gz',
     ...                         alphas=alphas_cues, betas=(0.1, 0.1),
     ...                         weights=weights)
 
@@ -231,7 +231,7 @@ If you prefer to get a ``xarray.DataArray`` returned you can set the flag ``make
 
 .. code-block:: python
 
-    >>> weights = ndl.dict_ndl(events='doc/data/lexample.tab.gz',
+    >>> weights = ndl.dict_ndl(events='docs/data/lexample.tab.gz',
     ...                        alphas=alphas_cues, betas=(0.1, 0.1),
     ...                        make_data_array=True)
     >>> weights  # doctest: +ELLIPSIS
@@ -275,8 +275,8 @@ generate an event file based on a raw corpus file and filter it:
 
     >>> import pyndl
     >>> from pyndl import preprocess
-    >>> preprocess.create_event_file(corpus_file='doc/data/lcorpus.txt',
-    ...                              event_file='doc/data/levent.tab.gz',
+    >>> preprocess.create_event_file(corpus_file='docs/data/lcorpus.txt',
+    ...                              event_file='docs/data/levent.tab.gz',
     ...                              allowed_symbols='a-zA-Z',
     ...                              context_structure='document',
     ...                              event_structure='consecutive_words',
[…]
