+++ doc/source/conf.py
@@ -67,6 +67,6 @@ html_static_path = ['_static']
 html_context = {"display_github": True,  # Add 'Edit on Github' link instead of 'View page source'
                 "github_user": 'quantling',
                 "github_repo": 'pyndl',
-                "github_version": 'master',
+                "github_version": 'main',
                 "conf_py_path": '/doc/source/'
                 }
