+++ doc/source/development.rst
@@ -1,9 +1,9 @@
 Development
 ===========
-.. image:: https://github.com/quantling/pyndl/actions/workflows/python-test.yml/badge.svg?branch=master
+.. image:: https://github.com/quantling/pyndl/actions/workflows/python-test.yml/badge.svg?branch=main
     :target: https://github.com/quantling/pyndl/actions/workflows/python-test.yml
 
-.. image:: https://codecov.io/gh/quantling/pyndl/branch/master/graph/badge.svg?token=2GWUXRA9PD
+.. image:: https://codecov.io/gh/quantling/pyndl/branch/main/graph/badge.svg?token=2GWUXRA9PD
     :target: https://codecov.io/gh/quantling/pyndl
 
 .. image:: https://img.shields.io/lgtm/grade/python/g/quantling/pyndl.svg?logo=lgtm&logoWidth=18
@@ -22,9 +22,9 @@ Getting Involved
 The *pyndl* project welcomes help in the following ways:
 
     * Making Pull Requests for
-      `code <https://github.com/quantling/pyndl/tree/master/pyndl>`_,
-      `tests <https://github.com/quantling/pyndl/tree/master/tests>`_
-      or `documentation <https://github.com/quantling/pyndl/tree/master/doc>`_.
+      `code <https://github.com/quantling/pyndl/tree/main/pyndl>`_,
+      `tests <https://github.com/quantling/pyndl/tree/main/tests>`_
+      or `documentation <https://github.com/quantling/pyndl/tree/main/doc>`_.
     * Commenting on `open issues <https://github.com/quantling/pyndl/issues>`_
       and `pull requests <https://github.com/quantling/pyndl/pulls>`_.
     * Helping to answer `questions in the issue section
@@ -167,7 +167,7 @@ is a good tool to do that. ``yep`` builds ontop of ``google-perftools``.
 (https://pypi.org/project/yep/)
 
 
-Keeping a fork in sync with master
+Keeping a fork in sync with main
 ----------------------------------
 
 .. note::
@@ -177,7 +177,7 @@ Keeping a fork in sync with master
     https://help.github.com/articles/working-with-forks/ .
 
 If you fork the ``pyndl`` project on github.com you might want to keep it in
-sync with master. In order to do so, you need to setup a remote repository
+sync with main. In order to do so, you need to setup a remote repository
 within a local ``pyndl`` clone of you fork. This remote repository will point
 to the original ``pyndl`` repository and is usually called ``upstream``. In
 order to do so run with a Terminal within the cloned pyndl folder:
@@ -193,12 +193,12 @@ local repository by running:
 
     git fetch upstream
 
-In order to sync you ``master`` branch run:
+In order to sync you ``main`` branch run:
 
 .. code:: bash
 
-    git checkout master
-    git merge upstream/master
+    git checkout main
+    git merge upstream/main
 
 If the merge cannot be fast-forward, you should resolve any issue now and
 commit the manually merged files.
@@ -257,23 +257,23 @@ Codecov          |codecov|                        Monitoring of test coverage
 LGTM             |lgtm|                           Monitoring code quality
 ===============  ===========  ==================  ===========================
 
-.. |actions| image:: https://github.com/quantling/pyndl/actions/workflows/python-test.yml/badge.svg?branch=master
+.. |actions| image:: https://github.com/quantling/pyndl/actions/workflows/python-test.yml/badge.svg?branch=main
     :target: https://github.com/quantling/pyndl/actions/workflows/python-test.yml
 
-.. |codecov| image:: https://codecov.io/gh/quantling/pyndl/branch/master/graph/badge.svg?token=2GWUXRA9PD
[…]
