### 2022-11-22 41731c3
renames default branch and docs folder (#237)
* Solves #234

* renames doc folder to docs

* changes default branch from master to main

* fixes all broken links from master to main and from doc to docs
