### 2022-11-22 ddf6e44
bumps seaborn version because of plot_result.py
