+++ pyproject.toml
@@ -45,7 +45,7 @@ flake8 = "^4.0.1"
 sphinx = "^1.4"
 sphinx_rtd_theme = "1.0.0"
 notebook = "^6.4.10"
-seaborn = "^0.11.2"
+seaborn = "^0.12.1"
 numpydoc = "1.2"
 easydev = "0.9.35"
 pylint = "^2.0.0"
