### 2022-11-22 389d6aa
increases dpi and adds background transparency
