+++ benchmark/plot_result.py
@@ -37,9 +37,9 @@ fg.legend.set_title(None)
 fg.axes[0, 0].set_title('Single processing')
 fg.axes[0, 1].set_title('Parallel processing (2 jobs)')
 fg.savefig('benchmark_result.pdf')         
-fg.savefig('benchmark_result.png')         
-fg.savefig('../paper/benchmark_result.png')         
-fg.savefig('../docs/source/_static/benchmark_result.png')         
+fg.savefig('benchmark_result.png', transparent=True, dpi=600)         
+fg.savefig('../paper/benchmark_result.png', transparent=True, dpi=600)         
+fg.savefig('../docs/source/_static/benchmark_result.png', transparent=True, dpi=600)         
 
 fg = sns.relplot(data=data_long.loc[['Rndl2-1thread', 'pyndl_openmp1'], :], kind="line", x="event_num", y="wctime", hue='lib', style='lib',
                  markers=True, dashes=False, errorbar='se')
@@ -48,7 +48,7 @@ plt.title('Single processing')
 plt.xlabel('events')
 plt.ylabel('wall-clock time [sec]')
 fg.savefig('single_result.pdf')
-fg.savefig('single_result.png')
+fg.savefig('single_result.png', transparent=True, dpi=600)
 
 fg = sns.relplot(data=data_long[data_long.parallel==True], kind="line", x="event_num", y="wctime", hue='lib', style='lib',
                  markers=True, dashes=False, errorbar='se', facet_kws=dict(legend_out=False))
@@ -57,4 +57,4 @@ plt.title('Parallel processing (2 jobs)')
 plt.xlabel('events')
 plt.ylabel('wall-clock time [sec]')
 fg.savefig('parallel_result.pdf')
-fg.savefig('parallel_result.png')
+fg.savefig('parallel_result.png', transparent=True, dpi=600)
