+++ .github/workflows/draft-pdf.yml
@@ -1,3 +1,5 @@
+name: Build paper PDF
+
 on: [push]
 
 jobs:
