### 2022-11-22 quantling/pyndl#228 @derNarr
We reformulated one paragraph in the paper to:

> The input to *pyndl* is agnostic to the actual domain as long as it is
> tokenized as Unicode character strings. Input sequences can consist of multiple
> tokens separated by underscores which is together with the tab-character the
> only special character in *pyndl*. While *pyndl* provides some basic
> preprocessing for grapheme tokenization, the preprocessing of ideograms,
> pictograms, logograms, and speech is possible using custom preprocessing. For
> example, word classification using tokenized speech audio recordings was
> investigated in @Arnold_2017. Inputs in this work consisted of around 50 tokens
> per time slice, where each token encoded the pitch, loudness, and variability
> into a string.

Hopefully, this makes the use-case for `pyndl` more clear and showcases what is possible and what is not.
