### 2022-11-23 780a691
adds sentence to thank reviewers and review process
