### 2022-11-23 #4515 @derNarr
Dear @osorensen, dear @frankier, dear @VenkteshV , dear @jinhangjiang ,

we have addressed now all of the urgent issues and improved the paper and the package documentation according to your comments. We especially oriented ourselves along this [comment](https://github.com/openjournals/joss-reviews/issues/4515#issuecomment-1201165829) .

The documentation of pyndl is substantially improved, we added a small benchmark and the code to run it, we made more clear that *pyndl* is mainly about a learning algorithm but that the Unicode support in Python3/Cython is a major advantage about the difficulties of properly processing Unicode in the R/C++ language. The claim of the "multi-language" support comes mainly from this fact, as it was for most quantitative linguists too painful to debug their code in R/C++ with ndl2. (Sometimes it worked flawlessly though even in R, which might have dependent on the installed locals of the machine.)

We added a citation to xarray and sanitized the paper to use US-style English and proper idiomatic expression. Furthermore, we clarified that the main focus is on grapheme clusters, but that other input formats even like speech audio recordings can be processed with *pyndl*, but only with custom preprocessing pipelines. We believe the same is true for ideograms / pictograms/ logograms.

In the acknowledgements, we added a sentence thanking the reviewers and acknowledging good review process (@frankier , @VenkteshV , @jinhangjiang  are you okay with this?):

> Finally, this paper and the associated package benefited from constructive and
> conscientious peer review. We would like to thank our three reviewers. Special
> thanks go to Frankie Robertson for his constructive and in-depth
> feedback and his suggestions how to make the package more user-friendly.

During our clean up, we renamed the `master` branch to `main` and renamed the folder containing the documentation from `doc` to `docs`. We published a new version of pyndl in version 1.0.0, which now supplies a wheel for amd64 Linux
systems.

How should we proceed now?

Best regards,

David-Elias, Marc, Elnaz, and Tino
