### 2022-11-23 2315404
adds basic workflow attempt to automatically upload package to PyPI
