+++ .github/workflows/python-publish.yml
@@ -1,0 +1,34 @@
+name: Test automatic upload
+
+on:
+  push:
+    branches:
+      - add-release-workflow
+  release:
+    types: [published]
+
+jobs:
+  deploy:
+    name: ${{ matrix.os }} ${{ matrix.python-version }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macOS-latest, windows-latest]
+        python-version: ['3.9']
+        poetry-version: [1.1.13]
+      fail-fast: false
+    runs-on: ${{ matrix.os }}
+
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Run image
+        uses: abatilo/actions-poetry@v2.1.3
+        with:
+          poetry-version: ${{ matrix.poetry-version }}
+      - name: Build and publish package
+        run:
+          - poetry config repositories.test-pypi https://test.pypi.org/legacy/
+          - poetry publish --build --username __token__ --password ${{ secrets.TEST_PYPI_API_TOKEN }} --repository test-pypi
