+++ .github/workflows/python-publish.yml
@@ -14,7 +14,7 @@ jobs:
       matrix:
         os: [ubuntu-latest, macOS-latest, windows-latest]
         python-version: ['3.9']
-        poetry-version: [1.1.13]
+        poetry-version: [1.2]
       fail-fast: false
     runs-on: ${{ matrix.os }}
 
@@ -33,4 +33,4 @@ jobs:
       - name: Install package
         run: poetry install
       - name: Build and publish package
-        run: poetry publish --build --username __token__ --password ${{ secrets.TEST_PYPI_API_TOKEN }} --repository test-pypi #--skip-existing
+        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.TEST_PYPI_API_TOKEN }} --repository test-pypi
