### 2022-11-23 5544889
drops windows and adds python 3.8-3.10
