+++ .github/workflows/python-publish.yml
@@ -12,8 +12,8 @@ jobs:
     name: ${{ matrix.os }} ${{ matrix.python-version }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macOS-latest, windows-latest]
-        python-version: ['3.9']
+        os: [ubuntu-latest, macOS-latest]
+        python-version: ['3.8', '3.9', '3.10']
         poetry-version: [1.2]
       fail-fast: false
     runs-on: ${{ matrix.os }}
