+++ .github/workflows/python-publish.yml
@@ -1,9 +1,6 @@
-name: Test automatic upload
+name: Build and upload releases
 
 on:
-  push:
-    branches:
-      - add-release-workflow
   release:
     types: [published]
 
@@ -28,9 +25,7 @@ jobs:
         uses: abatilo/actions-poetry@v2.1.3
         with:
           poetry-version: ${{ matrix.poetry-version }}
-      - name: Configure poetry
-        run: poetry config repositories.test-pypi https://test.pypi.org/legacy/
       - name: Install package
         run: poetry install
       - name: Build and publish package
-        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.TEST_PYPI_API_TOKEN }} --repository test-pypi
+        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.PYPI_API_TOKEN }}
