### 2022-11-24 b7dcceb
adds link to ndl2, which we finally got published to github
