+++ paper/paper.bib
@@ -66,6 +66,7 @@ https://doi.org/10.5334/jors.148
   author={Shaoul, Cyrus and Schilling, N and Bitschnau, S and Arppe, Antti and Hendrix, Peter and Baayen, R Harald},
   journal={R package version},
   volume={1},
+  url = {https://github.com/quantling/ndl2},
   year={2014}
 }
 
