### 2022-11-24 34c2e46
adds links to ndl2 into the README.rst as well; fixes #100
