+++ docs/source/credits.rst
@@ -58,15 +58,14 @@ If you are using BibTex you may want to use this example BibTex entry::
 
 Funding
 -------
-*pyndl* supported by the Alexander von Humboldt Professorship awarded to R.
-Harald Baayen, the ERC advanced Grant (no. 742545), and by the University of
-Tübingen.
+*pyndl* has been supported by the Alexander von Humboldt Professorship awarded
+to R.  Harald Baayen, the ERC advanced Grant WIDE (no. 742545), and the
+University of Tübingen.
 
 Acknowledgements
 ----------------
-This package is build as a python replacement for the R `ndl2 package
-<https://github.com/quantling/ndl2>`_. Some
-ideas on how to build the API and how to efficiently run the Rescorla Wagner
-iterative learning on large text corpora are inspired by the way the ndl2
-package solves this problems.
+This package is build as a Python replacement for the R `ndl2 package
+<https://github.com/quantling/ndl2>`_. Some ideas on how to build the API and
+how to efficiently run the Rescorla Wagner iterative learning on large text
+corpora are inspired by the ndl2 package.
 
