### 2022-11-24 #4515 @danielskatz
FYI, @editorialbot only sees a command if is the first thing in a comment, and thus also only one command can be in a comment
