### 2022-11-24 f14acd9
fixes errors under Windows by explicitly using utf-8 encodings when text files are read or written
