### 2022-11-25 #4515 @osorensen
Thanks @frankier, your thorough review is very much appreciated!

Regarding the paper length @derNarr, I think the paper is it is now is fine, because it seems like the large number of references inflates the word count a bit (and a large number of references is in general a good thing).
