### 2022-11-25 d9b245e
adds way to run test on non-Linux systems to documentation; bumps version to 1.1.0
