### 2022-11-25 ce10077
adds windows testing to workflow
