+++ .github/workflows/python-test.yml
@@ -13,7 +13,7 @@ jobs:
     name: ${{ matrix.os }} ${{ matrix.python-version }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macOS-latest]
+        os: [ubuntu-latest, macOS-latest, windows-latest]
         python-version: ['3.8', '3.10']
         poetry-version: [1.1.13]
       fail-fast: false
