### 2022-11-25 ecf512e
renames tests workflow to Tests to account for testing now also testing on Windows
