### 2022-11-25 c6c4479
fixes classifier to stable as package is no longer beta
