### 2022-11-25 6eabb03
bumps upper python version to 3.11 since it is released since 24-10-2022
