+++ .github/workflows/python-test.yml
@@ -14,7 +14,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macOS-latest, windows-latest]
-        python-version: ['3.8', '3.10']
+        python-version: ['3.8', '3.11']
         poetry-version: [1.1.13]
       fail-fast: false
     runs-on: ${{ matrix.os }}
