### 2022-11-25 2616013
adds Windows as supported operating system in pyproject.toml; adds wheel building for windows in Github workflows; adds python 3.11 for wheels
