+++ .github/workflows/python-publish.yml
@@ -9,8 +9,8 @@ jobs:
     name: ${{ matrix.os }} ${{ matrix.python-version }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macOS-latest]
-        python-version: ['3.8', '3.9', '3.10']
+        os: [ubuntu-latest, macOS-latest, windows-latest]
+        python-version: ['3.8', '3.9', '3.10', '3.11']
         poetry-version: [1.2]
       fail-fast: false
     runs-on: ${{ matrix.os }}
