+++ pyproject.toml
@@ -22,6 +22,7 @@ classifiers = ['Development Status :: 5 - Production/Stable',
                'Intended Audience :: Science/Research',
                'Operating System :: POSIX :: Linux',
                'Operating System :: MacOS',
+               'Operating System :: Microsoft :: Windows',
                'Topic :: Scientific/Engineering',
                'Topic :: Scientific/Engineering :: Artificial Intelligence',
                'Topic :: Scientific/Engineering :: Information Analysis',]
