### 2022-11-25 8765017
makes test for exception platform independent again
