### 2022-11-25 d4e80eb
adds testPyPI upload to test whether wheel build under windows works now
