+++ .github/workflows/python-publish.yml
@@ -3,6 +3,9 @@ name: Build and upload releases
 on:
   release:
     types: [published]
+  push:
+    branches:
+      - fix-windows-errors
 
 jobs:
   deploy:
@@ -25,7 +28,10 @@ jobs:
         uses: abatilo/actions-poetry@v2.1.3
         with:
           poetry-version: ${{ matrix.poetry-version }}
+      - name: Configure test PyPI
+        run: poetry config repositories.test-pypi https://test.pypi.org/legacy/
       - name: Install package
         run: poetry install
       - name: Build and publish package
-        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.PYPI_API_TOKEN }}
+        #run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.PYPI_API_TOKEN }}
+        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.TEST_PYPI_API_TOKEN }} --repository test-pypi
