### 2022-11-25 f76338d
resets release workflow and adds a workflow for testing automatic releases
