+++ .github/workflows/python-test-publish.yml
@@ -1,8 +1,9 @@
-name: Build and upload releases
+name: Test atomatic builds
 
 on:
-  release:
-    types: [published]
+  push:
+    branches:
+      - your-branch-name
 
 jobs:
   deploy:
@@ -25,7 +26,9 @@ jobs:
         uses: abatilo/actions-poetry@v2.1.3
         with:
           poetry-version: ${{ matrix.poetry-version }}
+      - name: Configure test PyPI
+        run: poetry config repositories.test-pypi https://test.pypi.org/legacy/
       - name: Install package
         run: poetry install
-      - name: Build and publish package
-        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.PYPI_API_TOKEN }}
+      - name: Build and upload package to tets PyPI
+        run: poetry publish --build --skip-existing --username __token__ --password ${{ secrets.TEST_PYPI_API_TOKEN }} --repository test-pypi
