### 2022-11-26 a075603
corrects license file; fixes #209
