+++ LICENSE.txt
@@ -1,4 +1,5 @@
-Copyright (c) 2017 David-Elias Künstle, Lennard Schneider, Konstantin Sering, Marc Weitz
+Copyright (c) 2017-2022 Konstantin Sering, Marc Weitz, David-Elias Künstle,
+Lennard Schneider, Elnaz Shafaei-Bejestan and open source contributers
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
