### 2022-11-26 40d6c91
switches easydev with sphinx-copybutton
