### 2022-11-26 c7faa80
removes automatic upload testing and sets package version to v1.1.0
