### 2022-11-26 8d3fd1e
fixes errors under Windows by explicitly using utf-8 encoding (#241)
* fixes errors under Windows by explicitly using utf-8 encodings when text files are read or written

* adds windows testing to workflow

* renames tests workflow to "Tests" to account for testing now also testing on Windows

* adds Windows as supported operating system in pyproject.toml

* adds wheel building for windows in Github workflows

* adds python 3.11 for wheels

* exchanges easydev with sphinx-copybutton

* adds way to run test on non-Linux systems to documentation

* changes installation description for Windows

*
[…]
