+++ .github/workflows/python-test.yml
@@ -1,4 +1,4 @@
-name: Linux & OSX tests
+name: Tests
 
 on:
   push:
@@ -13,16 +13,16 @@ jobs:
     name: ${{ matrix.os }} ${{ matrix.python-version }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macOS-latest]
-        python-version: ['3.8', '3.10']
+        os: [ubuntu-latest, macOS-latest, windows-latest]
+        python-version: ['3.8', '3.11']
         poetry-version: [1.1.13]
       fail-fast: false
     runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
 
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
 
@@ -47,7 +47,7 @@ jobs:
         run: poetry config virtualenvs.in-project true
 
       - name: Set up cache
-        uses: actions/cache@v2
+        uses: actions/cache@v3
         id: cache
         with:
           path: .venv
