+++ LICENSE.txt
@@ -1,5 +1,4 @@
-Copyright (c) 2017-2022 Konstantin Sering, Marc Weitz, David-Elias Künstle,
-Lennard Schneider, Elnaz Shafaei-Bejestan and open source contributers
+Copyright (c) 2017-2022 Konstantin Sering, Marc Weitz, David-Elias Künstle,Lennard Schneider, Elnaz Shafaei-Bejestan and open source contributers
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
