### 2022-11-26 9c333bf
removes old LICENSE file
