### 2022-11-26 57666f3
adds same syntax as in the other workflows
