### 2022-12-02 #4515 @osorensen
> Okay! I think the paper is probably a bit too long by JOSS standards now, which is 500-1000 words. (I think a little bit over 1000 words is okay, but it seems like it's probably too much now.) If it were possible to fit things on two pages that might be ideal. I suppose this is an editorial issue and so might be beyond my remit and not quite sure how much this is enforced.
>
> It's possible the plot could be made a bit smaller, which may help with reducing the number of pages (but not the number of words). It might be possible to fit it on a single pane, given that there need only be: ndl (1 job), ndl2 (1 job), ndl2 (2 jobs), pyndl (1 job), pyndl (2 jobs, openMP) pyndl (2 jobs, threading). You may need to move the legend outside the plotting area but it might still save a tiny bit of space.
>
> The acknowledgement is appreciated, but shortening/removing it helps you get to JOSS editorial standards it's fine by me.
>
> I agree that everything in my review has been addressed, and so I recommend that it be accepted, possibly pending some edits to the paper for length.

@frankier could you please tick off the final boxes in [your checklist](https://github.com/openjournals/joss-reviews/issues/4515#issuecomment-1174794445)?
