### 2022-12-02 #4515 @osorensen
👋 @VenkteshV, could you please let us know if the points raised in your review have been addressed, and if so, tick off the remaining points on [your checklist](https://github.com/openjournals/joss-reviews/issues/4515#issuecomment-1172759260)?
