### 2022-12-02 #4515 @VenkteshV
> wave @VenkteshV, could you please let us know if the points raised in your review have been addressed, and if so, tick off the remaining points on [your checklist](https://github.com/openjournals/joss-reviews/issues/4515#issuecomment-1172759260)?

Aplogies. Done!!
It is a fine module and apologies for delays from my end. I checked and my comments have been addressed. The documentation is more detailed and the writeup has also been changed incorporating the feedback. Thanks a lot. And aplogies for the delay. I was a bit occupied.
