### 2022-12-07 72724f3
Paper (#243)
* adds joss paper

* add github action and docker to build paper 
https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles

* matches titile in Zenodo to package paper title (correct upper casing)

* bumps version number to 1.1.1

Co-authored-by: Marc Weitz <marc.weitz@onlinehome.de>
Co-authored-by: David-Elias Künstle <dekuenstle@users.noreply.github.com>
Co-authored-by: Elnaz Shafaei <elnaz.shafaei@gmail.com>
