+++ paper/paper.bib
@@ -1,0 +1,233 @@
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {
+https://doi.org/10.5334/jors.148
+}
+}
+
+@incollection{rescorla1972theory,
+  title={A theory of Pavlovian conditioning: Variations in the effectiveness of reinforcement and nonreinforcement},
+  author={Rescorla, Robert A and Wagner, Allan R},
+  booktitle={Classical conditioning: Current research and theory},
+  year={1972}
+}
+
+@techreport{widrow1960adaptive,
+  title={Adaptive switching circuits},
+  author={Widrow, Bernard and Hoff, Marcian E},
+  year={1960},
+  institution={Stanford Univ Ca Stanford Electronics Labs},
+  doi = {10.21236/ad0241531}
+}
+
+@article{Danks_2003,
+	doi = {10.1016/s0022-2496(02)00016-0},
+	url = {https://doi.org/10.1016%2Fs0022-2496%2802%2900016-0},
+	year = 2003,
+	month = {apr},
+	publisher = {Elsevier {BV}},
+	volume = {47},
+	number = {2},
+	pages = {109--121},
+	author = {David Danks},
+	title = {Equilibria of the Rescorla{\textendash}Wagner model},
+	journal = {Journal of Mathematical Psychology}
+}
+
+@article{Baayen_2011,
+	doi = {10.1037/a0023851},
+	url = {https://doi.org/10.1037%2Fa0023851},
+	year = 2011,
+	month = {jul},
+	publisher = {American Psychological Association ({APA})},
+	volume = {118},
+	number = {3},
+	pages = {438--481},
+	author = {R. Harald Baayen and Petar Milin and Dušica Filipović Đurđević and Peter Hendrix and Marco Marelli},
+	title = {An amorphous model for morphological processing in visual comprehension based on naive discriminative learning.},
+	journal = {Psychological Review}
+}
+
+@article{ndl,
+  title={Package ‘ndl’},
+  author={Arppe, Antti and Hendrix, Peter and Milin, Petar and Baayen, R Harald and Sering, Tino and Shaoul, Cyrus},
+  year={2018}
+}
+
+@article{ndl2,
+  title={Ndl2: Na{\"\i}ve discriminative learning},
+  author={Shaoul, Cyrus and Schilling, N and Bitschnau, S and Arppe, Antti and Hendrix, Peter and Baayen, R Harald},
+  journal={R package version},
+  volume={1},
+  url = {https://github.com/quantling/ndl2},
+  year={2014}
+}
+
+@article{Shafaei_Bajestan_2021,
+	doi = {10.1080/23273798.2021.1954207},
+	url = {https://doi.org/10.1080%2F23273798.2021.1954207},
+	year = 2021,
+	month = {jul},
+	publisher = {Informa {UK} Limited},
+	pages = {1--28},
+	author = {Elnaz Shafaei-Bajestan and Masoumeh Moradipour-Tari and Peter Uhrig and R. Harald Baayen},
+	title = {{LDL}-{AURIS}: a computational model, grounded in error-driven learning, for the comprehension of single spoken words},
+	journal = {Language, Cognition and Neuroscience}
+}
+
+@article{TOMASCHEK_2019,
+	doi = {10.1017/s0022226719000203},
+	url = {https://doi.org/10.1017%2Fs0022226719000203},
+	year = 2019,
+	month = {jun},
[…]
