### 2022-12-07 #4515 @derNarr
* tagged release is this one here (tag `v1.1.1`): https://github.com/quantling/pyndl/releases/tag/v1.1.1
* Zenodo link https://zenodo.org/record/7409834
* title matches the paper title, authors are all relevant contributors, ordering in the paper and the package as documented on Zenodo differs on purpose
* Zenodo DOI for version pyndl 1.1.1 is: 10.5281/zenodo.7409834

Anything missing or incomplete?
