### 2022-12-15 #4515 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
